% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_operations.R
\name{billing_associate_source_views}
\alias{billing_associate_source_views}
\title{Associates one or more source billing views with an existing billing
view}
\usage{
billing_associate_source_views(arn, sourceViews)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the billing view to associate source
views with.}

\item{sourceViews}{[required] A list of ARNs of the source billing views to associate.}
}
\description{
Associates one or more source billing views with an existing billing view. This allows creating aggregate billing views that combine data from multiple sources.

See \url{https://www.paws-r-sdk.com/docs/billing_associate_source_views/} for full documentation.
}
\keyword{internal}
