% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_operations.R
\name{billing_get_billing_view}
\alias{billing_get_billing_view}
\title{Returns the metadata associated to the specified billing view ARN}
\usage{
billing_get_billing_view(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) that can be used to uniquely identify the
billing view.}
}
\description{
Returns the metadata associated to the specified billing view ARN.

See \url{https://www.paws-r-sdk.com/docs/billing_get_billing_view/} for full documentation.
}
\keyword{internal}
