% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_operations.R
\name{billing_get_enterprise_support_charge_summary}
\alias{billing_get_enterprise_support_charge_summary}
\title{Returns a summary of Enterprise Support data aggregated across all
accounts in the Enterprise Support profile}
\usage{
billing_get_enterprise_support_charge_summary(billingMonth)
}
\arguments{
\item{billingMonth}{[required] The billing month in YYYY-MM format. This must be a month in the past.}
}
\description{
Returns a summary of Enterprise Support data aggregated across all accounts in the Enterprise Support profile.

See \url{https://www.paws-r-sdk.com/docs/billing_get_enterprise_support_charge_summary/} for full documentation.
}
\keyword{internal}
