% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_operations.R
\name{billing_get_resource_policy}
\alias{billing_get_resource_policy}
\title{Returns the resource-based policy document attached to the resource in
JSON format}
\usage{
billing_get_resource_policy(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the billing view resource to which the
policy is attached to.}
}
\description{
Returns the resource-based policy document attached to the resource in \code{JSON} format.

See \url{https://www.paws-r-sdk.com/docs/billing_get_resource_policy/} for full documentation.
}
\keyword{internal}
