% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_operations.R
\name{billing_list_billing_view_segments}
\alias{billing_list_billing_view_segments}
\title{Lists the segments of a billing view over a given time period}
\usage{
billing_list_billing_view_segments(
  timeRange = NULL,
  arn = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{timeRange}{The billing period to query. If you don't provide a time range, the current billing period, which is the calendar month in UTC, is used.}

\item{arn}{The Amazon Resource Name (ARN) that uniquely identifies the billing view to query. If you don't provide an ARN, the caller's \code{PRIMARY} billing view is used. The ARN must reference a primary billing view. Custom billing views aren't supported.}

\item{maxResults}{The number of entries a paginated response contains. Valid values range from 1 to 100. The default is 100.}

\item{nextToken}{The pagination token that is used on subsequent calls to list billing view segments.}
}
\description{
Lists the segments of a billing view over a given time period. Each segment identifies the billing domain (\code{PRO_FORMA} or \code{BILLABLE}) and the account relationships that apply during its time range.

See \url{https://www.paws-r-sdk.com/docs/billing_list_billing_view_segments/} for full documentation.
}
\keyword{internal}
