% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_operations.R
\name{billing_list_business_support_account_charges}
\alias{billing_list_business_support_account_charges}
\title{Returns Business Support charges broken down at the linked account level
for a given billing month}
\usage{
billing_list_business_support_account_charges(
  billingMonth,
  accountId = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{billingMonth}{[required] The billing month to retrieve Business Support charges for, in YYYY-MM format. You can request the current month (charges will be estimated) or a past month (charges will be finalized).}

\item{accountId}{The linked account ID to filter results to a specific account. If you don't specify a value, the response includes charges for all linked accounts.}

\item{maxResults}{The maximum number of results to return per page. Default is 100.}

\item{nextToken}{The pagination token for the next page of results.}
}
\description{
Returns Business Support charges broken down at the linked account level for a given billing month.

See \url{https://www.paws-r-sdk.com/docs/billing_list_business_support_account_charges/} for full documentation.
}
\keyword{internal}
