% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_operations.R
\name{billing_list_source_views_for_billing_view}
\alias{billing_list_source_views_for_billing_view}
\title{Lists the source views (managed Amazon Web Services billing views)
associated with the billing view}
\usage{
billing_list_source_views_for_billing_view(
  arn,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) that can be used to uniquely identify the
billing view.}

\item{maxResults}{The number of entries a paginated response contains.}

\item{nextToken}{The pagination token that is used on subsequent calls to list billing
views.}
}
\description{
Lists the source views (managed Amazon Web Services billing views) associated with the billing view.

See \url{https://www.paws-r-sdk.com/docs/billing_list_source_views_for_billing_view/} for full documentation.
}
\keyword{internal}
