% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_operations.R
\name{billing_list_tags_for_resource}
\alias{billing_list_tags_for_resource}
\title{Lists tags associated with the billing view resource}
\usage{
billing_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Lists tags associated with the billing view resource.

See \url{https://www.paws-r-sdk.com/docs/billing_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
