% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_operations.R
\name{billing_redeem_credits}
\alias{billing_redeem_credits}
\title{Redeems an Amazon Web Services promotional credit code on behalf of the
calling account}
\usage{
billing_redeem_credits(promoCode)
}
\arguments{
\item{promoCode}{[required] The promotional credit code to redeem.}
}
\description{
Redeems an Amazon Web Services promotional credit code on behalf of the calling account. On success, a new credit is added to the account's credit ledger with the amount, validity period, and applicable products defined by the promotion. The credit is then automatically applied to subsequent bills according to the standard credit application order.

See \url{https://www.paws-r-sdk.com/docs/billing_redeem_credits/} for full documentation.
}
\keyword{internal}
