% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_operations.R
\name{billing_tag_resource}
\alias{billing_tag_resource}
\title{An API operation for adding one or more tags (key-value pairs) to a
resource}
\usage{
billing_tag_resource(resourceArn, resourceTags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{resourceTags}{[required] A list of tag key value pairs that are associated with the resource.}
}
\description{
An API operation for adding one or more tags (key-value pairs) to a resource.

See \url{https://www.paws-r-sdk.com/docs/billing_tag_resource/} for full documentation.
}
\keyword{internal}
