% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_operations.R
\name{billing_update_billing_view}
\alias{billing_update_billing_view}
\title{An API to update the attributes of the billing view}
\usage{
billing_update_billing_view(
  arn,
  name = NULL,
  description = NULL,
  dataFilterExpression = NULL
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) that can be used to uniquely identify the
billing view.}

\item{name}{The name of the billing view.}

\item{description}{The description of the billing view.}

\item{dataFilterExpression}{See
\href{https://docs.aws.amazon.com/aws-cost-management/latest/APIReference/API_billing_Expression.html}{Expression}.
Billing view only supports \code{LINKED_ACCOUNT} and \code{Tags}.}
}
\description{
An API to update the attributes of the billing view.

See \url{https://www.paws-r-sdk.com/docs/billing_update_billing_view/} for full documentation.
}
\keyword{internal}
