% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_associate_accounts}
\alias{billingconductor_associate_accounts}
\title{Connects an array of account IDs in a consolidated billing family to a
predefined billing group}
\usage{
billingconductor_associate_accounts(Arn, AccountIds)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the billing group that associates the
array of account IDs.}

\item{AccountIds}{[required] The associating array of account IDs.}
}
\description{
Connects an array of account IDs in a consolidated billing family to a predefined billing group. The account IDs must be a part of the consolidated billing family during the current month, and not already associated with another billing group. The maximum number of accounts that can be associated in one call is 30.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_associate_accounts/} for full documentation.
}
\keyword{internal}
