% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_associate_pricing_rules}
\alias{billingconductor_associate_pricing_rules}
\title{Connects an array of PricingRuleArns to a defined PricingPlan}
\usage{
billingconductor_associate_pricing_rules(Arn, PricingRuleArns)
}
\arguments{
\item{Arn}{[required] The \code{PricingPlanArn} that the \code{PricingRuleArns} are associated with.}

\item{PricingRuleArns}{[required] The \code{PricingRuleArns} that are associated with the Pricing Plan.}
}
\description{
Connects an array of \code{PricingRuleArns} to a defined \code{PricingPlan}. The maximum number \code{PricingRuleArn} that can be associated in one call is 30.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_associate_pricing_rules/} for full documentation.
}
\keyword{internal}
