% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_delete_pricing_plan}
\alias{billingconductor_delete_pricing_plan}
\title{Deletes a pricing plan}
\usage{
billingconductor_delete_pricing_plan(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the pricing plan that you're deleting.}
}
\description{
Deletes a pricing plan. The pricing plan must not be associated with any billing groups to delete successfully.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_delete_pricing_plan/} for full documentation.
}
\keyword{internal}
