% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_get_billing_transfer_preference}
\alias{billingconductor_get_billing_transfer_preference}
\title{Retrieves the auto billing group creation preference for a billing
transfer}
\usage{
billingconductor_get_billing_transfer_preference(ResponsibilityTransferArn)
}
\arguments{
\item{ResponsibilityTransferArn}{[required] The Amazon Resource Name (ARN) of the billing transfer whose preference you want to retrieve.}
}
\description{
Retrieves the auto billing group creation preference for a billing transfer.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_get_billing_transfer_preference/} for full documentation.
}
\keyword{internal}
