% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_list_billing_group_cost_reports}
\alias{billingconductor_list_billing_group_cost_reports}
\title{A paginated call to retrieve a summary report of actual Amazon Web
Services charges and the calculated Amazon Web Services charges based on
the associated pricing plan of a billing group}
\usage{
billingconductor_list_billing_group_cost_reports(
  BillingPeriod = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  Filters = NULL
)
}
\arguments{
\item{BillingPeriod}{The preferred billing period for your report.}

\item{MaxResults}{The maximum number of reports to retrieve.}

\item{NextToken}{The pagination token used on subsequent calls to get reports.}

\item{Filters}{A \code{ListBillingGroupCostReportsFilter} to specify billing groups to
retrieve reports from.}
}
\description{
A paginated call to retrieve a summary report of actual Amazon Web Services charges and the calculated Amazon Web Services charges based on the associated pricing plan of a billing group.

See \url{https://paws-r.github.io/docs/billingconductor/list_billing_group_cost_reports.html} for full documentation.
}
\keyword{internal}
