% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_list_billing_groups}
\alias{billingconductor_list_billing_groups}
\title{A paginated call to retrieve a list of billing groups for the given
billing period}
\usage{
billingconductor_list_billing_groups(
  BillingPeriod = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  Filters = NULL
)
}
\arguments{
\item{BillingPeriod}{The preferred billing period to get billing groups.}

\item{MaxResults}{The maximum number of billing groups to retrieve.}

\item{NextToken}{The pagination token that's used on subsequent calls to get billing
groups.}

\item{Filters}{A \code{ListBillingGroupsFilter} that specifies the billing group and pricing
plan to retrieve billing group information.}
}
\description{
A paginated call to retrieve a list of billing groups for the given billing period. If you don't provide a billing group, the current billing period is used.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_list_billing_groups/} for full documentation.
}
\keyword{internal}
