% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_list_pricing_plans_associated_with_pricing_rule}
\alias{billingconductor_list_pricing_plans_associated_with_pricing_rule}
\title{A list of the pricing plans that are associated with a pricing rule}
\usage{
billingconductor_list_pricing_plans_associated_with_pricing_rule(
  BillingPeriod = NULL,
  PricingRuleArn,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{BillingPeriod}{The pricing plan billing period for which associations will be listed.}

\item{PricingRuleArn}{[required] The pricing rule Amazon Resource Name (ARN) for which associations will
be listed.}

\item{MaxResults}{The optional maximum number of pricing rule associations to retrieve.}

\item{NextToken}{The optional pagination token returned by a previous call.}
}
\description{
A list of the pricing plans that are associated with a pricing rule.

See \url{https://paws-r.github.io/docs/billingconductor/list_pricing_plans_associated_with_pricing_rule.html} for full documentation.
}
\keyword{internal}
