% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_list_pricing_rules}
\alias{billingconductor_list_pricing_rules}
\title{Describes a pricing rule that can be associated to a pricing plan, or
set of pricing plans}
\usage{
billingconductor_list_pricing_rules(
  BillingPeriod = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{BillingPeriod}{The preferred billing period to get the pricing plan.}

\item{Filters}{A \code{DescribePricingRuleFilter} that specifies the Amazon Resource Name
(ARNs) of pricing rules to retrieve pricing rules information.}

\item{MaxResults}{The maximum number of pricing rules to retrieve.}

\item{NextToken}{The pagination token that's used on subsequent call to get pricing
rules.}
}
\description{
Describes a pricing rule that can be associated to a pricing plan, or set of pricing plans.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_list_pricing_rules/} for full documentation.
}
\keyword{internal}
