% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_list_resources_associated_to_custom_line_item}
\alias{billingconductor_list_resources_associated_to_custom_line_item}
\title{List the resources that are associated to a custom line item}
\usage{
billingconductor_list_resources_associated_to_custom_line_item(
  BillingPeriod = NULL,
  Arn,
  MaxResults = NULL,
  NextToken = NULL,
  Filters = NULL
)
}
\arguments{
\item{BillingPeriod}{The billing period for which the resource associations will be listed.}

\item{Arn}{[required] The ARN of the custom line item for which the resource associations will
be listed.}

\item{MaxResults}{(Optional) The maximum number of resource associations to be retrieved.}

\item{NextToken}{(Optional) The pagination token that's returned by a previous request.}

\item{Filters}{(Optional) A \code{ListResourcesAssociatedToCustomLineItemFilter} that can
specify the types of resources that should be retrieved.}
}
\description{
List the resources that are associated to a custom line item.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_list_resources_associated_to_custom_line_item/} for full documentation.
}
\keyword{internal}
