% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_list_tags_for_resource}
\alias{billingconductor_list_tags_for_resource}
\title{A list the tags for a resource}
\usage{
billingconductor_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource to list the
tags.}
}
\description{
A list the tags for a resource.

See \url{https://paws-r.github.io/docs/billingconductor/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
