% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_tag_resource}
\alias{billingconductor_tag_resource}
\title{Associates the specified tags to a resource with the specified
resourceArn}
\usage{
billingconductor_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to which to add tags.}

\item{Tags}{[required] The tags to add to the resource as a list of key-value pairs.}
}
\description{
Associates the specified tags to a resource with the specified \code{resourceArn}. If existing tags on a resource are not specified in the request parameters, they are not changed.

See \url{https://paws-r.github.io/docs/billingconductor/tag_resource.html} for full documentation.
}
\keyword{internal}
