% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_untag_resource}
\alias{billingconductor_untag_resource}
\title{Deletes specified tags from a resource}
\usage{
billingconductor_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to which to delete tags.}

\item{TagKeys}{[required] The tags to delete from the resource as a list of key-value pairs.}
}
\description{
Deletes specified tags from a resource.

See \url{https://paws-r.github.io/docs/billingconductor/untag_resource.html} for full documentation.
}
\keyword{internal}
