% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_update_billing_group}
\alias{billingconductor_update_billing_group}
\title{This updates an existing billing group}
\usage{
billingconductor_update_billing_group(
  Arn,
  Name = NULL,
  Status = NULL,
  ComputationPreference = NULL,
  Description = NULL
)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the billing group being updated.}

\item{Name}{The name of the billing group. The names must be unique to each billing
group.}

\item{Status}{The status of the billing group. Only one of the valid values can be
used.}

\item{ComputationPreference}{The preferences and settings that will be used to compute the Amazon Web
Services charges for a billing group.}

\item{Description}{A description of the billing group.}
}
\description{
This updates an existing billing group.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_update_billing_group/} for full documentation.
}
\keyword{internal}
