% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_update_billing_transfer_preference}
\alias{billingconductor_update_billing_transfer_preference}
\title{Sets the auto billing group creation preference for a billing transfer}
\usage{
billingconductor_update_billing_transfer_preference(
  ClientToken = NULL,
  ResponsibilityTransferArn,
  AutoBillingTransferBillingGroupCreation
)
}
\arguments{
\item{ClientToken}{A unique, case-sensitive identifier that you specify to ensure idempotency of the request. Idempotency ensures that an API request completes no more than one time. With an idempotent request, if the original request completes successfully, any subsequent retries complete successfully without performing any further actions.}

\item{ResponsibilityTransferArn}{[required] The Amazon Resource Name (ARN) of the billing transfer whose preference you want to set.}

\item{AutoBillingTransferBillingGroupCreation}{[required] The auto billing group creation preference to set for the billing transfer.}
}
\description{
Sets the auto billing group creation preference for a billing transfer. When the preference is enabled, Billing Conductor automatically creates an indirect billing transfer billing group in your account, with the pricing plan that you specify, for each account that transfers its bill to the bill source account of this billing transfer. The preference applies only to billing groups that are created after you enable it.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_update_billing_transfer_preference/} for full documentation.
}
\keyword{internal}
