% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_update_pricing_plan}
\alias{billingconductor_update_pricing_plan}
\title{This updates an existing pricing plan}
\usage{
billingconductor_update_pricing_plan(Arn, Name = NULL, Description = NULL)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the pricing plan that you're updating.}

\item{Name}{The name of the pricing plan. The name must be unique to each pricing
plan.}

\item{Description}{The description of the pricing plan.}
}
\description{
This updates an existing pricing plan.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_update_pricing_plan/} for full documentation.
}
\keyword{internal}
