% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_create_notification}
\alias{budgets_create_notification}
\title{Creates a notification}
\usage{
budgets_create_notification(AccountId, BudgetName, Notification, Subscribers)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget that you want to
create a notification for.}

\item{BudgetName}{[required] The name of the budget that you want Amazon Web Services to notify you
about. Budget names must be unique within an account.}

\item{Notification}{[required] The notification that you want to create.}

\item{Subscribers}{[required] A list of subscribers that you want to associate with the notification.
Each notification can have one SNS subscriber and up to 10 email
subscribers.}
}
\description{
Creates a notification. You must create the budget before you create the associated notification.

See \url{https://www.paws-r-sdk.com/docs/budgets_create_notification/} for full documentation.
}
\keyword{internal}
