% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_create_subscriber}
\alias{budgets_create_subscriber}
\title{Creates a subscriber}
\usage{
budgets_create_subscriber(AccountId, BudgetName, Notification, Subscriber)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget that you want to
create a subscriber for.}

\item{BudgetName}{[required] The name of the budget that you want to subscribe to. Budget names must
be unique within an account.}

\item{Notification}{[required] The notification that you want to create a subscriber for.}

\item{Subscriber}{[required] The subscriber that you want to associate with a budget notification.}
}
\description{
Creates a subscriber. You must create the associated budget and notification before you create the subscriber.

See \url{https://www.paws-r-sdk.com/docs/budgets_create_subscriber/} for full documentation.
}
\keyword{internal}
