% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_delete_budget}
\alias{budgets_delete_budget}
\title{Deletes a budget}
\usage{
budgets_delete_budget(AccountId, BudgetName)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget that you want to
delete.}

\item{BudgetName}{[required] The name of the budget that you want to delete.}
}
\description{
Deletes a budget. You can delete your budget at any time.

See \url{https://paws-r.github.io/docs/budgets/delete_budget.html} for full documentation.
}
\keyword{internal}
