% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget_action}
\alias{budgets_describe_budget_action}
\title{Describes a budget action detail}
\usage{
budgets_describe_budget_action(AccountId, BudgetName, ActionId)
}
\arguments{
\item{AccountId}{[required] The account ID of the user. It's a 12-digit number.}

\item{BudgetName}{[required] A string that represents the budget name. The ":" and "\\" characters, and the "/action/" substring, aren't allowed.

Budget names are validated for content. Names that contain phone numbers, URLs, or email addresses combined with certain terms may be rejected.}

\item{ActionId}{[required] A system-generated universally unique identifier (UUID) for the action.}
}
\description{
Describes a budget action detail.

See \url{https://www.paws-r-sdk.com/docs/budgets_describe_budget_action/} for full documentation.
}
\keyword{internal}
