% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget_action_histories}
\alias{budgets_describe_budget_action_histories}
\title{Describes a budget action history detail}
\usage{
budgets_describe_budget_action_histories(
  AccountId,
  BudgetName,
  ActionId,
  TimePeriod = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AccountId}{[required] The account ID of the user. It's a 12-digit number.}

\item{BudgetName}{[required] A string that represents the budget name. The ":" and "\\" characters, and the "/action/" substring, aren't allowed.

Budget names are validated for content. Names that contain phone numbers, URLs, or email addresses combined with certain terms may be rejected.}

\item{ActionId}{[required] A system-generated universally unique identifier (UUID) for the action.}

\item{TimePeriod}{The period of time that's covered by a budget. The period has a start date and an end date. The start date must come before the end date. There are no restrictions on the end date.}

\item{MaxResults}{An integer that represents how many entries a paginated response contains. The maximum is 100.}

\item{NextToken}{A generic string.}
}
\description{
Describes a budget action history detail.

See \url{https://www.paws-r-sdk.com/docs/budgets_describe_budget_action_histories/} for full documentation.
}
\keyword{internal}
