% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget_actions_for_account}
\alias{budgets_describe_budget_actions_for_account}
\title{Describes all of the budget actions for an account}
\usage{
budgets_describe_budget_actions_for_account(
  AccountId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AccountId}{[required] The account ID of the user. It's a 12-digit number.}

\item{MaxResults}{An integer that represents how many entries a paginated response contains. The maximum is 100.}

\item{NextToken}{A generic string.}
}
\description{
Describes all of the budget actions for an account.

See \url{https://www.paws-r-sdk.com/docs/budgets_describe_budget_actions_for_account/} for full documentation.
}
\keyword{internal}
