% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget_actions_for_budget}
\alias{budgets_describe_budget_actions_for_budget}
\title{Describes all of the budget actions for a budget}
\usage{
budgets_describe_budget_actions_for_budget(
  AccountId,
  BudgetName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AccountId}{[required] The account ID of the user. It's a 12-digit number.}

\item{BudgetName}{[required] A string that represents the budget name. The ":" and "\\" characters, and the "/action/" substring, aren't allowed.

Budget names are validated for content. Names that contain phone numbers, URLs, or email addresses combined with certain terms may be rejected.}

\item{MaxResults}{An integer that represents how many entries a paginated response contains. The maximum is 100.}

\item{NextToken}{A generic string.}
}
\description{
Describes all of the budget actions for a budget.

See \url{https://www.paws-r-sdk.com/docs/budgets_describe_budget_actions_for_budget/} for full documentation.
}
\keyword{internal}
