% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget_performance_history}
\alias{budgets_describe_budget_performance_history}
\title{Describes the history for DAILY, MONTHLY, and QUARTERLY budgets}
\usage{
budgets_describe_budget_performance_history(
  AccountId,
  BudgetName,
  TimePeriod = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AccountId}{[required] The account ID of the user. It's a 12-digit number.}

\item{BudgetName}{[required] A string that represents the budget name. The ":" and "\\" characters, and the "/action/" substring, aren't allowed.

Budget names are validated for content. Names that contain phone numbers, URLs, or email addresses combined with certain terms may be rejected.}

\item{TimePeriod}{Retrieves how often the budget went into an \code{ALARM} state for the specified time period.}

\item{MaxResults}{An integer that represents how many entries a paginated response contains. The maximum is 100.}

\item{NextToken}{A generic string.}
}
\description{
Describes the history for \code{DAILY}, \code{MONTHLY}, and \code{QUARTERLY} budgets. Budget history isn't available for \code{ANNUAL} budgets.

See \url{https://www.paws-r-sdk.com/docs/budgets_describe_budget_performance_history/} for full documentation.
}
\keyword{internal}
