% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_list_tags_for_resource}
\alias{budgets_list_tags_for_resource}
\title{Lists tags associated with a budget or budget action resource}
\usage{
budgets_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The unique identifier for the resource.}
}
\description{
Lists tags associated with a budget or budget action resource.

See \url{https://www.paws-r-sdk.com/docs/budgets_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
