% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_tag_resource}
\alias{budgets_tag_resource}
\title{Creates tags for a budget or budget action resource}
\usage{
budgets_tag_resource(ResourceARN, ResourceTags)
}
\arguments{
\item{ResourceARN}{[required] The unique identifier for the resource.}

\item{ResourceTags}{[required] The tags associated with the resource.}
}
\description{
Creates tags for a budget or budget action resource.

See \url{https://www.paws-r-sdk.com/docs/budgets_tag_resource/} for full documentation.
}
\keyword{internal}
