% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_update_budget}
\alias{budgets_update_budget}
\title{Updates a budget}
\usage{
budgets_update_budget(AccountId, NewBudget)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget that you want to
update.}

\item{NewBudget}{[required] The budget that you want to update your budget to.}
}
\description{
Updates a budget. You can change every part of a budget except for the \code{budgetName} and the \code{calculatedSpend}. When you modify a budget, the \code{calculatedSpend} drops to zero until Amazon Web Services has new usage data to use for forecasting.

See \url{https://paws-r.github.io/docs/budgets/update_budget.html} for full documentation.
}
\keyword{internal}
