% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_update_budget_action}
\alias{budgets_update_budget_action}
\title{Updates a budget action}
\usage{
budgets_update_budget_action(
  AccountId,
  BudgetName,
  ActionId,
  NotificationType = NULL,
  ActionThreshold = NULL,
  Definition = NULL,
  ExecutionRoleArn = NULL,
  ApprovalModel = NULL,
  Subscribers = NULL
)
}
\arguments{
\item{AccountId}{[required] The account ID of the user. It's a 12-digit number.}

\item{BudgetName}{[required] A string that represents the budget name. The ":" and "\\" characters, and the "/action/" substring, aren't allowed.

Budget names are validated for content. Names that contain phone numbers, URLs, or email addresses combined with certain terms may be rejected.}

\item{ActionId}{[required] A system-generated universally unique identifier (UUID) for the action.}

\item{NotificationType}{The type of a notification. It must be ACTUAL or FORECASTED.}

\item{ActionThreshold}{The trigger threshold of the action.}

\item{Definition}{Specifies all of the type-specific parameters.}

\item{ExecutionRoleArn}{The role passed for action execution and reversion. Roles and actions must be in the same account.}

\item{ApprovalModel}{This specifies if the action needs manual or automatic approval.}

\item{Subscribers}{A list of subscribers.}
}
\description{
Updates a budget action.

See \url{https://www.paws-r-sdk.com/docs/budgets_update_budget_action/} for full documentation.
}
\keyword{internal}
