% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_update_subscriber}
\alias{budgets_update_subscriber}
\title{Updates a subscriber}
\usage{
budgets_update_subscriber(
  AccountId,
  BudgetName,
  Notification,
  OldSubscriber,
  NewSubscriber
)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget whose subscriber you
want to update.}

\item{BudgetName}{[required] The name of the budget whose subscriber you want to update.}

\item{Notification}{[required] The notification whose subscriber you want to update.}

\item{OldSubscriber}{[required] The previous subscriber that is associated with a budget notification.}

\item{NewSubscriber}{[required] The updated subscriber that is associated with a budget notification.}
}
\description{
Updates a subscriber.

See \url{https://www.paws-r-sdk.com/docs/budgets_update_subscriber/} for full documentation.
}
\keyword{internal}
