% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_delete_report_definition}
\alias{costandusagereportservice_delete_report_definition}
\title{Deletes the specified report}
\usage{
costandusagereportservice_delete_report_definition(ReportName)
}
\arguments{
\item{ReportName}{The name of the report that you want to delete. The name must be unique,
is case sensitive, and can't include spaces.}
}
\description{
Deletes the specified report.
}
\section{Request syntax}{
\preformatted{svc$delete_report_definition(
  ReportName = "string"
)
}
}

\examples{
\dontrun{
# The following example deletes the AWS Cost and Usage report named
# ExampleReport.
svc$delete_report_definition(
  ReportName = "ExampleReport"
)
}

}
\keyword{internal}
