% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_describe_report_definitions}
\alias{costandusagereportservice_describe_report_definitions}
\title{Lists the AWS Cost and Usage reports available to this account}
\usage{
costandusagereportservice_describe_report_definitions(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Lists the AWS Cost and Usage reports available to this account.

See \url{https://paws-r.github.io/docs/costandusagereportservice/describe_report_definitions.html} for full documentation.
}
\keyword{internal}
