% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_list_tags_for_resource}
\alias{costandusagereportservice_list_tags_for_resource}
\title{Lists the tags associated with the specified report definition}
\usage{
costandusagereportservice_list_tags_for_resource(ReportName)
}
\arguments{
\item{ReportName}{[required] The report name of the report definition that tags are to be returned
for.}
}
\description{
Lists the tags associated with the specified report definition.

See \url{https://www.paws-r-sdk.com/docs/costandusagereportservice_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
