% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_modify_report_definition}
\alias{costandusagereportservice_modify_report_definition}
\title{Allows you to programmatically update your report preferences}
\usage{
costandusagereportservice_modify_report_definition(
  ReportName,
  ReportDefinition
)
}
\arguments{
\item{ReportName}{[required] The name of the report that you want to create. The name must be unique, is case sensitive, and can't include spaces.}

\item{ReportDefinition}{[required] The definition of Amazon Web Services Cost and Usage Report. You can specify the report name, time unit, report format, compression format, S3 bucket, additional artifacts, and schema elements in the definition.}
}
\description{
Allows you to programmatically update your report preferences.

See \url{https://www.paws-r-sdk.com/docs/costandusagereportservice_modify_report_definition/} for full documentation.
}
\keyword{internal}
