% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_tag_resource}
\alias{costandusagereportservice_tag_resource}
\title{Associates a set of tags with a report definition}
\usage{
costandusagereportservice_tag_resource(ReportName, Tags)
}
\arguments{
\item{ReportName}{[required] The report name of the report definition that tags are to be associated
with.}

\item{Tags}{[required] The tags to be assigned to the report definition resource.}
}
\description{
Associates a set of tags with a report definition.

See \url{https://www.paws-r-sdk.com/docs/costandusagereportservice_tag_resource/} for full documentation.
}
\keyword{internal}
