% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_delete_anomaly_monitor}
\alias{costexplorer_delete_anomaly_monitor}
\title{Deletes a cost anomaly monitor}
\usage{
costexplorer_delete_anomaly_monitor(MonitorArn)
}
\arguments{
\item{MonitorArn}{[required] The unique identifier of the cost anomaly monitor that you want to
delete.}
}
\value{
An empty list.
}
\description{
Deletes a cost anomaly monitor.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_anomaly_monitor(
  MonitorArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
