% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_delete_cost_category_definition}
\alias{costexplorer_delete_cost_category_definition}
\title{Deletes a Cost Category}
\usage{
costexplorer_delete_cost_category_definition(CostCategoryArn)
}
\arguments{
\item{CostCategoryArn}{[required] The unique identifier for your Cost Category.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CostCategoryArn = "string",
  EffectiveEnd = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a Cost Category. Expenses from this month going forward will no
longer be categorized with this Cost Category.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_cost_category_definition(
  CostCategoryArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
