% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_describe_cost_category_definition}
\alias{costexplorer_describe_cost_category_definition}
\title{Returns the name, Amazon Resource Name (ARN), rules, definition, and
effective dates of a Cost Category that's defined in the account}
\usage{
costexplorer_describe_cost_category_definition(
  CostCategoryArn,
  EffectiveOn = NULL
)
}
\arguments{
\item{CostCategoryArn}{[required] The unique identifier for your Cost Category.}

\item{EffectiveOn}{The date when the Cost Category was effective.}
}
\description{
Returns the name, Amazon Resource Name (ARN), rules, definition, and effective dates of a Cost Category that's defined in the account.

See \url{https://paws-r.github.io/docs/costexplorer/describe_cost_category_definition.html} for full documentation.
}
\keyword{internal}
