% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_get_approximate_usage_records}
\alias{costexplorer_get_approximate_usage_records}
\title{Retrieves estimated usage records for hourly granularity or
resource-level data at daily granularity}
\usage{
costexplorer_get_approximate_usage_records(
  Granularity,
  Services = NULL,
  ApproximationDimension
)
}
\arguments{
\item{Granularity}{[required] How granular you want the data to be. You can enable data at hourly or
daily granularity.}

\item{Services}{The service metadata for the service or services you want to query. If
not specified, all elements are returned.}

\item{ApproximationDimension}{[required] The service to evaluate for the usage records. You can choose
resource-level data at daily granularity, or hourly granularity with or
without resource-level data.}
}
\description{
Retrieves estimated usage records for hourly granularity or resource-level data at daily granularity.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_get_approximate_usage_records/} for full documentation.
}
\keyword{internal}
