% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_get_rightsizing_recommendation}
\alias{costexplorer_get_rightsizing_recommendation}
\title{Creates recommendations that help you save cost by identifying idle and
underutilized Amazon EC2 instances}
\usage{
costexplorer_get_rightsizing_recommendation(
  Filter = NULL,
  Configuration = NULL,
  Service,
  PageSize = NULL,
  NextPageToken = NULL
)
}
\arguments{
\item{Filter}{Use \code{Expression} to filter in various Cost Explorer APIs.

Not all \code{Expression} types are supported in each API. Refer to the documentation for each specific API to see what is supported.

There are two patterns:
\itemize{
\item Simple dimension values.
\itemize{
\item There are three types of simple dimension values: \code{CostCategories}, \code{Tags}, and \code{Dimensions}.
\itemize{
\item Specify the \code{CostCategories} field to define a filter that acts on Cost Categories.
\item Specify the \code{Tags} field to define a filter that acts on Cost Allocation Tags.
\item Specify the \code{Dimensions} field to define a filter that acts on the \href{https://docs.aws.amazon.com/aws-cost-management/latest/APIReference/API_DimensionValues.html}{\code{DimensionValues}} .
}
\item For each filter type, you can set the dimension name and values for the filters that you plan to use.
\itemize{
\item For example, you can filter for \verb{REGION==us-east-1 OR REGION==us-west-1}. For \code{\link[=costexplorer_get_rightsizing_recommendation]{get_rightsizing_recommendation}}, the Region is a full name (for example, \verb{REGION==US East (N. Virginia)}.
\item The corresponding \code{Expression} for this example is as follows: \verb{\{ "Dimensions": \{ "Key": "REGION", "Values": [ "us-east-1", "us-west-1" ] \} \}}
\item As shown in the previous example, lists of dimension values are combined with \code{OR} when applying the filter.
}
\item You can also set different match options to further control how the filter behaves. Not all APIs support match options. Refer to the documentation for each specific API to see what is supported.
\itemize{
\item For example, you can filter for linked account names that start with "a".
\item The corresponding \code{Expression} for this example is as follows: \verb{\{ "Dimensions": \{ "Key": "LINKED_ACCOUNT_NAME", "MatchOptions": [ "STARTS_WITH" ], "Values": [ "a" ] \} \}}
}
}
\item Compound \code{Expression} types with logical operations.
\itemize{
\item You can use multiple \code{Expression} types and the logical operators \code{AND/OR/NOT} to create a list of one or more \code{Expression} objects. By doing this, you can filter by more advanced options.
\item For example, you can filter by \verb{((REGION == us-east-1 OR REGION == us-west-1) OR (TAG.Type == Type1)) AND (USAGE_TYPE != DataTransfer)}.
\item The corresponding \code{Expression} for this example is as follows: \verb{\{ "And": [ \{"Or": [ \{"Dimensions": \{ "Key": "REGION", "Values": [ "us-east-1", "us-west-1" ] \}\}, \{"Tags": \{ "Key": "TagName", "Values": ["Value1"] \} \} ]\}, \{"Not": \{"Dimensions": \{ "Key": "USAGE_TYPE", "Values": ["DataTransfer"] \}\}\} ] \} }
}

Because each \code{Expression} can have only one operator, the service returns an error if more than one is specified. The following example shows an \code{Expression} object that creates an error: \verb{\{ "And": [ ... ], "Dimensions": \{ "Key": "USAGE_TYPE", "Values": [ "DataTransfer" ] \} \}}

The following is an example of the corresponding error message: \code{"Expression has more than one roots. Only one root operator is allowed for each expression: And, Or, Not, Dimensions, Tags, CostCategories"}
}

For the \code{\link[=costexplorer_get_rightsizing_recommendation]{get_rightsizing_recommendation}} action, a combination of OR and NOT isn't supported. OR isn't supported between different dimensions, or dimensions and tags. NOT operators aren't supported. Dimensions are also limited to \code{LINKED_ACCOUNT}, \code{REGION}, or \code{RIGHTSIZING_TYPE}.

For the \code{\link[=costexplorer_get_reservation_purchase_recommendation]{get_reservation_purchase_recommendation}} action, only NOT is supported. AND and OR aren't supported. Dimensions are limited to \code{LINKED_ACCOUNT}.}

\item{Configuration}{You can use Configuration to customize recommendations across two attributes. You can choose to view recommendations for instances within the same instance families or across different instance families. You can also choose to view your estimated savings that are associated with recommendations with consideration of existing Savings Plans or RI benefits, or neither.}

\item{Service}{[required] The specific service that you want recommendations for. The only valid value for \code{\link[=costexplorer_get_rightsizing_recommendation]{get_rightsizing_recommendation}} is "\code{AmazonEC2}".}

\item{PageSize}{The number of recommendations that you want returned in a single response object.}

\item{NextPageToken}{The pagination token that indicates the next set of results that you want to retrieve.}
}
\description{
Creates recommendations that help you save cost by identifying idle and underutilized Amazon EC2 instances.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_get_rightsizing_recommendation/} for full documentation.
}
\keyword{internal}
