% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_get_savings_plan_purchase_recommendation_details}
\alias{costexplorer_get_savings_plan_purchase_recommendation_details}
\title{Retrieves the details for a Savings Plan recommendation}
\usage{
costexplorer_get_savings_plan_purchase_recommendation_details(
  RecommendationDetailId
)
}
\arguments{
\item{RecommendationDetailId}{[required] The ID that is associated with the Savings Plan recommendation.}
}
\description{
Retrieves the details for a Savings Plan recommendation. These details include the hourly data-points that construct the cost, coverage, and utilization charts.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_get_savings_plan_purchase_recommendation_details/} for full documentation.
}
\keyword{internal}
