% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_list_cost_allocation_tag_backfill_history}
\alias{costexplorer_list_cost_allocation_tag_backfill_history}
\title{Retrieves a list of your historical cost allocation tag backfill
requests}
\usage{
costexplorer_list_cost_allocation_tag_backfill_history(
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{NextToken}{The token to retrieve the next set of results. Amazon Web Services
provides the token when the response from a previous call has more
results than the maximum page size.}

\item{MaxResults}{The maximum number of objects that are returned for this request.}
}
\description{
Retrieves a list of your historical cost allocation tag backfill requests.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_list_cost_allocation_tag_backfill_history/} for full documentation.
}
\keyword{internal}
