% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_list_cost_allocation_tags}
\alias{costexplorer_list_cost_allocation_tags}
\title{Get a list of cost allocation tags}
\usage{
costexplorer_list_cost_allocation_tags(
  Status = NULL,
  TagKeys = NULL,
  Type = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Status}{The status of cost allocation tag keys that are returned for this
request.}

\item{TagKeys}{The list of cost allocation tag keys that are returned for this request.}

\item{Type}{The type of \code{CostAllocationTag} object that are returned for this
request. The \code{AWSGenerated} type tags are tags that Amazon Web Services
defines and applies to support Amazon Web Services resources for cost
allocation purposes. The \code{UserDefined} type tags are tags that you
define, create, and apply to resources.}

\item{NextToken}{The token to retrieve the next set of results. Amazon Web Services
provides the token when the response from a previous call has more
results than the maximum page size.}

\item{MaxResults}{The maximum number of objects that are returned for this request. By
default, the request returns 100 results.}
}
\description{
Get a list of cost allocation tags. All inputs in the API are optional and serve as filters. By default, all cost allocation tags are returned.

See \url{https://paws-r.github.io/docs/costexplorer/list_cost_allocation_tags.html} for full documentation.
}
\keyword{internal}
