% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_list_cost_category_resource_associations}
\alias{costexplorer_list_cost_category_resource_associations}
\title{Returns resource associations of all cost categories defined in the
account}
\usage{
costexplorer_list_cost_category_resource_associations(
  CostCategoryArn = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{CostCategoryArn}{The unique identifier for your cost category.}

\item{NextToken}{The token to retrieve the next set of results. Amazon Web Services
provides the token when the response from a previous call has more
results than the maximum page size.}

\item{MaxResults}{The number of entries a paginated response contains.}
}
\description{
Returns resource associations of all cost categories defined in the account. You have the option to use \code{CostCategoryArn} to get the association for a specific cost category. \code{\link[=costexplorer_list_cost_category_resource_associations]{list_cost_category_resource_associations}} supports pagination. The request can have a \code{MaxResults} range up to 100.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_list_cost_category_resource_associations/} for full documentation.
}
\keyword{internal}
