% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_list_savings_plans_purchase_recommendation_generation}
\alias{costexplorer_list_savings_plans_purchase_recommendation_generation}
\title{Retrieves a list of your historical recommendation generations within
the past 30 days}
\usage{
costexplorer_list_savings_plans_purchase_recommendation_generation(
  GenerationStatus = NULL,
  RecommendationIds = NULL,
  PageSize = NULL,
  NextPageToken = NULL
)
}
\arguments{
\item{GenerationStatus}{The status of the recommendation generation.}

\item{RecommendationIds}{The IDs for each specific recommendation.}

\item{PageSize}{The number of recommendations that you want returned in a single
response object.}

\item{NextPageToken}{The token to retrieve the next set of results.}
}
\description{
Retrieves a list of your historical recommendation generations within the past 30 days.

See \url{https://paws-r.github.io/docs/costexplorer/list_savings_plans_purchase_recommendation_generation.html} for full documentation.
}
\keyword{internal}
