% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_list_tags_for_resource}
\alias{costexplorer_list_tags_for_resource}
\title{Returns a list of resource tags associated with the resource specified
by the Amazon Resource Name (ARN)}
\usage{
costexplorer_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource. For a list of supported
resources, see
\href{https://docs.aws.amazon.com/aws-cost-management/latest/APIReference/API_ResourceTag.html}{ResourceTag}.}
}
\description{
Returns a list of resource tags associated with the resource specified by the Amazon Resource Name (ARN).

See \url{https://paws-r.github.io/docs/costexplorer/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
