% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_provide_anomaly_feedback}
\alias{costexplorer_provide_anomaly_feedback}
\title{Modifies the feedback property of a given cost anomaly}
\usage{
costexplorer_provide_anomaly_feedback(AnomalyId, Feedback)
}
\arguments{
\item{AnomalyId}{[required] A cost anomaly ID.}

\item{Feedback}{[required] Describes whether the cost anomaly was a planned activity or you
considered it an anomaly.}
}
\description{
Modifies the feedback property of a given cost anomaly.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_provide_anomaly_feedback/} for full documentation.
}
\keyword{internal}
