% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_start_commitment_purchase_analysis}
\alias{costexplorer_start_commitment_purchase_analysis}
\title{Specifies the parameters of a planned commitment purchase and starts the
generation of the analysis}
\usage{
costexplorer_start_commitment_purchase_analysis(
  CommitmentPurchaseAnalysisConfiguration
)
}
\arguments{
\item{CommitmentPurchaseAnalysisConfiguration}{[required] The configuration for the commitment purchase analysis.}
}
\description{
Specifies the parameters of a planned commitment purchase and starts the generation of the analysis. This enables you to estimate the cost, coverage, and utilization impact of your planned commitment purchases.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_start_commitment_purchase_analysis/} for full documentation.
}
\keyword{internal}
