% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_start_cost_allocation_tag_backfill}
\alias{costexplorer_start_cost_allocation_tag_backfill}
\title{Request a cost allocation tag backfill}
\usage{
costexplorer_start_cost_allocation_tag_backfill(BackfillFrom)
}
\arguments{
\item{BackfillFrom}{[required] The date you want the backfill to start from. The date can only be a
first day of the month (a billing start date). Dates can't precede the
previous twelve months, or in the future.}
}
\description{
Request a cost allocation tag backfill. This will backfill the activation status (either \code{active} or \code{inactive}) for all tag keys from \code{para:BackfillFrom} up to the time this request is made.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_start_cost_allocation_tag_backfill/} for full documentation.
}
\keyword{internal}
