% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_untag_resource}
\alias{costexplorer_untag_resource}
\title{Removes one or more tags from a resource}
\usage{
costexplorer_untag_resource(ResourceArn, ResourceTagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource. For a list of supported
resources, see
\href{https://docs.aws.amazon.com/aws-cost-management/latest/APIReference/API_ResourceTag.html}{ResourceTag}.}

\item{ResourceTagKeys}{[required] A list of tag keys associated with tags that need to be removed from the
resource. If you specify a tag key that doesn't exist, it's ignored.
Although the maximum number of array members is 200, user-tag maximum is
50. The remaining are reserved for Amazon Web Services use.}
}
\description{
Removes one or more tags from a resource. Specify only tag keys in your request. Don't specify the value.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_untag_resource/} for full documentation.
}
\keyword{internal}
