% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_update_anomaly_monitor}
\alias{costexplorer_update_anomaly_monitor}
\title{Updates an existing cost anomaly monitor}
\usage{
costexplorer_update_anomaly_monitor(MonitorArn, MonitorName)
}
\arguments{
\item{MonitorArn}{[required] Cost anomaly monitor Amazon Resource Names (ARNs).}

\item{MonitorName}{The new name for the cost anomaly monitor.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  MonitorArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates an existing cost anomaly monitor. The changes made are applied
going forward, and does not change anomalies detected in the past.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_anomaly_monitor(
  MonitorArn = "string",
  MonitorName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
