% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_update_anomaly_subscription}
\alias{costexplorer_update_anomaly_subscription}
\title{Updates an existing cost anomaly monitor subscription}
\usage{
costexplorer_update_anomaly_subscription(
  SubscriptionArn,
  Threshold = NULL,
  Frequency = NULL,
  MonitorArnList = NULL,
  Subscribers = NULL,
  SubscriptionName = NULL
)
}
\arguments{
\item{SubscriptionArn}{[required] A cost anomaly subscription Amazon Resource Name (ARN).}

\item{Threshold}{The update to the threshold value for receiving notifications.}

\item{Frequency}{The update to the frequency value that subscribers receive
notifications.}

\item{MonitorArnList}{A list of cost anomaly monitor ARNs.}

\item{Subscribers}{The update to the subscriber list.}

\item{SubscriptionName}{The new name of the subscription.}
}
\description{
Updates an existing cost anomaly monitor subscription.

See \url{https://paws-r.github.io/docs/costexplorer/update_anomaly_subscription.html} for full documentation.
}
\keyword{internal}
