% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_batch_describe_entities}
\alias{marketplacecatalog_batch_describe_entities}
\title{Returns metadata and content for multiple entities}
\usage{
marketplacecatalog_batch_describe_entities(EntityRequestList)
}
\arguments{
\item{EntityRequestList}{[required] List of entity IDs and the catalogs the entities are present in.}
}
\description{
Returns metadata and content for multiple entities.

See \url{https://www.paws-r-sdk.com/docs/marketplacecatalog_batch_describe_entities/} for full documentation.
}
\keyword{internal}
