% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_cancel_change_set}
\alias{marketplacecatalog_cancel_change_set}
\title{Used to cancel an open change request}
\usage{
marketplacecatalog_cancel_change_set(Catalog, ChangeSetId)
}
\arguments{
\item{Catalog}{[required] Required. The catalog related to the request. Fixed value:
\code{AWSMarketplace}.}

\item{ChangeSetId}{[required] Required. The unique identifier of the
\code{\link[=marketplacecatalog_start_change_set]{start_change_set}} request that
you want to cancel.}
}
\description{
Used to cancel an open change request. Must be sent before the status of the request changes to \code{APPLYING}, the final stage of completing your change request. You can describe a change during the 60-day request history retention period for API calls.

See \url{https://paws-r.github.io/docs/marketplacecatalog/cancel_change_set.html} for full documentation.
}
\keyword{internal}
