% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_delete_resource_policy}
\alias{marketplacecatalog_delete_resource_policy}
\title{Deletes a resource-based policy on an Entity that is identified by its
resource ARN}
\usage{
marketplacecatalog_delete_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Entity resource that is associated
with the resource policy.}
}
\description{
Deletes a resource-based policy on an Entity that is identified by its resource ARN.

See \url{https://paws-r.github.io/docs/marketplacecatalog/delete_resource_policy.html} for full documentation.
}
\keyword{internal}
