% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_get_resource_policy}
\alias{marketplacecatalog_get_resource_policy}
\title{Gets a resource-based policy of an Entity that is identified by its
resource ARN}
\usage{
marketplacecatalog_get_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Entity resource that is associated
with the resource policy.}
}
\description{
Gets a resource-based policy of an Entity that is identified by its resource ARN.

See \url{https://paws-r.github.io/docs/marketplacecatalog/get_resource_policy.html} for full documentation.
}
\keyword{internal}
