% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_list_assessments}
\alias{marketplacecatalog_list_assessments}
\title{Returns a paginated list of assessments associated with an entity or
change set in AWS Marketplace}
\usage{
marketplacecatalog_list_assessments(
  Catalog,
  FrameworkId = NULL,
  AssessmentTargetFilter = NULL,
  FrameworkFilters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Catalog}{[required] The catalog related to the request. Fixed value: \code{AWSMarketplace}}

\item{FrameworkId}{The unique identifier of a framework. When specified, only assessments performed against this framework are returned. For example, \code{AMISecurity}.}

\item{AssessmentTargetFilter}{Filters the list of assessments to those performed against a specific entity or change set.}

\item{FrameworkFilters}{Framework-specific filters. Set exactly one member to filter results to assessments performed against that framework.}

\item{MaxResults}{Specifies the upper limit of the elements on a single page. If a value isn't provided, the default value is 20. Valid values range from 1 to 100.}

\item{NextToken}{The value of the next token, if it exists. \code{null} if there are no more results.}
}
\description{
Returns a paginated list of assessments associated with an entity or change set in AWS Marketplace. An \emph{assessment} is the result of evaluating a product or change set against a framework, such as AMI Security or Container Security.

See \url{https://www.paws-r-sdk.com/docs/marketplacecatalog_list_assessments/} for full documentation.
}
\keyword{internal}
