% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_put_resource_policy}
\alias{marketplacecatalog_put_resource_policy}
\title{Attaches a resource-based policy to an Entity}
\usage{
marketplacecatalog_put_resource_policy(ResourceArn, Policy)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the Entity resource you want to
associate with a resource policy.}

\item{Policy}{[required] The policy document to set; formatted in JSON.}
}
\description{
Attaches a resource-based policy to an Entity. Examples of an entity include: \code{AmiProduct} and \code{ContainerProduct}.

See \url{https://paws-r.github.io/docs/marketplacecatalog/put_resource_policy.html} for full documentation.
}
\keyword{internal}
