% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacemetering_operations.R
\name{marketplacemetering_batch_meter_usage}
\alias{marketplacemetering_batch_meter_usage}
\title{Amazon Web Services Marketplace is introducing Concurrent Agreements,
enabling buyers to make multiple purchases per Amazon Web Services
account}
\usage{
marketplacemetering_batch_meter_usage(UsageRecords, ProductCode = NULL)
}
\arguments{
\item{UsageRecords}{[required] The set of \code{UsageRecords} to submit. \code{\link[=marketplacemetering_batch_meter_usage]{batch_meter_usage}} accepts up to 25 \code{UsageRecords} at a time.}

\item{ProductCode}{Product code is used to uniquely identify a product in Amazon Web Services Marketplace. The product code should be the same as the one used during the publishing of a new product.

\code{ProductCode} is required only for legacy integrations that use \code{CustomerIdentifier}. For new integrations using \code{LicenseArn} (Concurrent Agreements), do NOT include \code{ProductCode} at the request level. The \code{LicenseArn} in each \code{UsageRecord} identifies both the product and the specific agreement.

Sending metering records with both \code{ProductCode} and \code{LicenseArn} for the same customer within the same hour will result in duplicate billing. If you are migrating from product-based metering to license-based metering, stop sending \code{ProductCode} before you start sending \code{LicenseArn}.}
}
\description{
Amazon Web Services Marketplace is introducing Concurrent Agreements, enabling buyers to make multiple purchases per Amazon Web Services account. Starting June 1, 2026, new SaaS products must use \code{CustomerAWSAccountId} (instead of \code{CustomerIdentifier}), \code{LicenseArn} (instead of \code{ProductCode}) to support this feature. \code{\link[=marketplacemetering_batch_meter_usage]{batch_meter_usage}} does not support \code{CustomerIdentifier} for new integrations. Existing integrations continue to work. Review the new integration for Concurrent Agreements \href{https://catalog.workshops.aws/mpseller/en-US/saas/integration-for-concurrent-agreements}{here}. For additional implementation details, see \href{https://docs.aws.amazon.com/marketplace/latest/userguide/saas-code-examples.html#saas-batchmeterusage-licensearn-example}{BatchMeterUsage code example with LicenseArn} in the \emph{Amazon Web Services Marketplace Seller Guide}.

See \url{https://www.paws-r-sdk.com/docs/marketplacemetering_batch_meter_usage/} for full documentation.
}
\keyword{internal}
