% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacemetering_operations.R
\name{marketplacemetering_resolve_customer}
\alias{marketplacemetering_resolve_customer}
\title{ResolveCustomer is called by a SaaS application during the registration
process}
\usage{
marketplacemetering_resolve_customer(RegistrationToken)
}
\arguments{
\item{RegistrationToken}{[required] When a buyer visits your website during the registration process, the buyer submits a registration token through the browser. The registration token is resolved to obtain a \code{CustomerIdentifier} along with the \code{CustomerAWSAccountId}, \code{ProductCode}, and \code{LicenseArn}.

For new SaaS product integrations, the \code{CustomerIdentifier} field is not populated. Use \code{CustomerAWSAccountId} and \code{LicenseArn} for customer identification.}
}
\description{
\code{\link[=marketplacemetering_resolve_customer]{resolve_customer}} is called by a SaaS application during the registration process. When a buyer visits your website during the registration process, the buyer submits a registration token through their browser. The registration token is resolved through this API to obtain a \code{CustomerIdentifier} along with the \code{CustomerAWSAccountId}, \code{ProductCode}, and \code{LicenseArn}.

See \url{https://www.paws-r-sdk.com/docs/marketplacemetering_resolve_customer/} for full documentation.
}
\keyword{internal}
