% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_delete_alias}
\alias{paymentcryptographycontrolplane_delete_alias}
\title{Deletes the alias, but doesn't affect the underlying key}
\usage{
paymentcryptographycontrolplane_delete_alias(AliasName)
}
\arguments{
\item{AliasName}{[required] A friendly name that you can use to refer Amazon Web Services Payment
Cryptography key. This value must begin with \verb{alias/} followed by a
name, such as \code{alias/ExampleAlias}.}
}
\description{
Deletes the alias, but doesn't affect the underlying key.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_delete_alias/} for full documentation.
}
\keyword{internal}
