% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_disable_default_key_replication_regions}
\alias{paymentcryptographycontrolplane_disable_default_key_replication_regions}
\title{Disables Multi-Region key replication settings for the specified Amazon
Web Services Regions in your Amazon Web Services account, preventing new
keys from being automatically replicated to those regions}
\usage{
paymentcryptographycontrolplane_disable_default_key_replication_regions(
  ReplicationRegions
)
}
\arguments{
\item{ReplicationRegions}{[required] The list of Amazon Web Services Regions to remove from the account's
default replication regions.

New keys created after this operation will not automatically be
replicated to these regions, though existing keys with replication to
these regions will be unaffected.}
}
\description{
Disables \href{https://docs.aws.amazon.com/payment-cryptography/latest/userguide/keys-multi-region-replication.html}{Multi-Region key replication} settings for the specified Amazon Web Services Regions in your Amazon Web Services account, preventing new keys from being automatically replicated to those regions.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_disable_default_key_replication_regions/} for full documentation.
}
\keyword{internal}
