% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_disassociate_mpa_team}
\alias{paymentcryptographycontrolplane_disassociate_mpa_team}
\title{Removes the association between a Multi-Party Approval (MPA) team and a
protected operation}
\usage{
paymentcryptographycontrolplane_disassociate_mpa_team(
  Action,
  RequesterComment = NULL
)
}
\arguments{
\item{Action}{[required] The protected operation to disassociate from the MPA team. Currently,
the only supported value is \code{IMPORT_ROOT_PUBLIC_KEY_CERTIFICATE}.}

\item{RequesterComment}{The comment from the requester explaining the reason for the
disassociation.

Don't include personal, confidential or sensitive information in this
field. This field may be displayed in plaintext in CloudTrail logs and
other output.}
}
\description{
Removes the association between a Multi-Party Approval (MPA) team and a protected operation.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_disassociate_mpa_team/} for full documentation.
}
\keyword{internal}
