% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_enable_default_key_replication_regions}
\alias{paymentcryptographycontrolplane_enable_default_key_replication_regions}
\title{Enables Multi-Region key replication settings for your Amazon Web
Services account, causing new keys to be automatically replicated to the
specified Amazon Web Services Regions when created}
\usage{
paymentcryptographycontrolplane_enable_default_key_replication_regions(
  ReplicationRegions
)
}
\arguments{
\item{ReplicationRegions}{[required] The list of Amazon Web Services Regions to enable as default replication
regions for the Amazon Web Services account for \href{https://docs.aws.amazon.com/payment-cryptography/latest/userguide/keys-multi-region-replication.html}{Multi-Region key replication}.

New keys created in this account will automatically be replicated to
these regions unless explicitly overridden during key creation.}
}
\description{
Enables \href{https://docs.aws.amazon.com/payment-cryptography/latest/userguide/keys-multi-region-replication.html}{Multi-Region key replication} settings for your Amazon Web Services account, causing new keys to be automatically replicated to the specified Amazon Web Services Regions when created.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_enable_default_key_replication_regions/} for full documentation.
}
\keyword{internal}
