% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_get_alias}
\alias{paymentcryptographycontrolplane_get_alias}
\title{Gets the Amazon Web Services Payment Cryptography key associated with
the alias}
\usage{
paymentcryptographycontrolplane_get_alias(AliasName)
}
\arguments{
\item{AliasName}{[required] The alias of the Amazon Web Services Payment Cryptography key.}
}
\description{
Gets the Amazon Web Services Payment Cryptography key associated with the alias.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_get_alias/} for full documentation.
}
\keyword{internal}
