% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_get_certificate_signing_request}
\alias{paymentcryptographycontrolplane_get_certificate_signing_request}
\title{Creates a certificate signing request (CSR) from a key pair}
\usage{
paymentcryptographycontrolplane_get_certificate_signing_request(
  KeyIdentifier,
  SigningAlgorithm,
  CertificateSubject
)
}
\arguments{
\item{KeyIdentifier}{[required] Asymmetric key used for generating the certificate signing request}

\item{SigningAlgorithm}{[required] The cryptographic algorithm used to sign your CSR.}

\item{CertificateSubject}{[required] The metadata used to create the CSR.}
}
\description{
Creates a certificate signing request (CSR) from a key pair.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_get_certificate_signing_request/} for full documentation.
}
\keyword{internal}
