% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_get_default_key_replication_regions}
\alias{paymentcryptographycontrolplane_get_default_key_replication_regions}
\title{Retrieves the list of Amazon Web Services Regions where Multi-Region key
replication is currently enabled for your Amazon Web Services account}
\usage{
paymentcryptographycontrolplane_get_default_key_replication_regions()
}
\description{
Retrieves the list of Amazon Web Services Regions where \href{https://docs.aws.amazon.com/payment-cryptography/latest/userguide/keys-multi-region-replication.html}{Multi-Region key replication} is currently enabled for your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_get_default_key_replication_regions/} for full documentation.
}
\keyword{internal}
