% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_get_key}
\alias{paymentcryptographycontrolplane_get_key}
\title{Gets the key material for an Amazon Web Services Payment Cryptography
key, including the immutable and mutable data specified when the key was
created}
\usage{
paymentcryptographycontrolplane_get_key(KeyIdentifier)
}
\arguments{
\item{KeyIdentifier}{[required] The \code{KeyARN} of the Amazon Web Services Payment Cryptography key.}
}
\description{
Gets the key material for an Amazon Web Services Payment Cryptography key, including the immutable and mutable data specified when the key was created.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_get_key/} for full documentation.
}
\keyword{internal}
