% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_get_mpa_team_association}
\alias{paymentcryptographycontrolplane_get_mpa_team_association}
\title{Returns the Multi-Party Approval (MPA) team association for a protected
operation}
\usage{
paymentcryptographycontrolplane_get_mpa_team_association(Action)
}
\arguments{
\item{Action}{[required] The protected operation whose MPA team association you want to retrieve.
Currently, the only supported value is
\code{IMPORT_ROOT_PUBLIC_KEY_CERTIFICATE}.}
}
\description{
Returns the Multi-Party Approval (MPA) team association for a protected operation.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_get_mpa_team_association/} for full documentation.
}
\keyword{internal}
