% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_get_public_key_certificate}
\alias{paymentcryptographycontrolplane_get_public_key_certificate}
\title{Gets the public key certificate of the asymmetric key pair that exists
within Amazon Web Services Payment Cryptography}
\usage{
paymentcryptographycontrolplane_get_public_key_certificate(KeyIdentifier)
}
\arguments{
\item{KeyIdentifier}{[required] The \code{KeyARN} of the asymmetric key pair.}
}
\description{
Gets the public key certificate of the asymmetric key pair that exists within Amazon Web Services Payment Cryptography.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_get_public_key_certificate/} for full documentation.
}
\keyword{internal}
