% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_get_resource_policy}
\alias{paymentcryptographycontrolplane_get_resource_policy}
\title{Returns the resource-based policy attached to an Amazon Web Services
Payment Cryptography key}
\usage{
paymentcryptographycontrolplane_get_resource_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The \code{KeyARN} of the key whose resource-based policy you want to
retrieve.}
}
\description{
Returns the resource-based policy attached to an Amazon Web Services Payment Cryptography key.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_get_resource_policy/} for full documentation.
}
\keyword{internal}
