% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_list_tags_for_resource}
\alias{paymentcryptographycontrolplane_list_tags_for_resource}
\title{Lists the tags for an Amazon Web Services resource}
\usage{
paymentcryptographycontrolplane_list_tags_for_resource(
  MaxResults = NULL,
  NextToken = NULL,
  ResourceArn
)
}
\arguments{
\item{MaxResults}{Use this parameter to specify the maximum number of items to return.
When this value is present, Amazon Web Services Payment Cryptography
does not return more than the specified number of items, but it might
return fewer.

This value is optional. If you include a value, it must be between 1 and
100, inclusive. If you do not include a value, it defaults to 50.}

\item{NextToken}{Use this parameter in a subsequent request after you receive a response
with truncated results. Set it to the value of \code{NextToken} from the
truncated response you just received.}

\item{ResourceArn}{[required] The \code{KeyARN} of the key whose tags you are getting.}
}
\description{
Lists the tags for an Amazon Web Services resource.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
