% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_put_resource_policy}
\alias{paymentcryptographycontrolplane_put_resource_policy}
\title{Attaches or replaces a resource-based policy on an Amazon Web Services
Payment Cryptography key}
\usage{
paymentcryptographycontrolplane_put_resource_policy(ResourceArn, Policy)
}
\arguments{
\item{ResourceArn}{[required] The \code{KeyARN} of the key to attach the resource-based policy to.}

\item{Policy}{[required] The resource-based policy to attach to the key, in JSON format.}
}
\description{
Attaches or replaces a resource-based policy on an Amazon Web Services Payment Cryptography key. A resource-based policy can grant cross-account access to your key.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_put_resource_policy/} for full documentation.
}
\keyword{internal}
