% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_restore_key}
\alias{paymentcryptographycontrolplane_restore_key}
\title{Cancels a scheduled key deletion during the waiting period}
\usage{
paymentcryptographycontrolplane_restore_key(KeyIdentifier)
}
\arguments{
\item{KeyIdentifier}{[required] The \code{KeyARN} of the key to be restored within Amazon Web Services
Payment Cryptography.}
}
\description{
Cancels a scheduled key deletion during the waiting period. Use this operation to restore a \code{Key} that is scheduled for deletion.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_restore_key/} for full documentation.
}
\keyword{internal}
