% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_start_key_usage}
\alias{paymentcryptographycontrolplane_start_key_usage}
\title{Enables an Amazon Web Services Payment Cryptography key, which makes it
active for cryptographic operations within Amazon Web Services Payment
Cryptography}
\usage{
paymentcryptographycontrolplane_start_key_usage(KeyIdentifier)
}
\arguments{
\item{KeyIdentifier}{[required] The \code{KeyArn} of the key.}
}
\description{
Enables an Amazon Web Services Payment Cryptography key, which makes it active for cryptographic operations within Amazon Web Services Payment Cryptography

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_start_key_usage/} for full documentation.
}
\keyword{internal}
