% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_tag_resource}
\alias{paymentcryptographycontrolplane_tag_resource}
\title{Adds or edits tags on an Amazon Web Services Payment Cryptography key}
\usage{
paymentcryptographycontrolplane_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The \code{KeyARN} of the key whose tags are being updated.}

\item{Tags}{[required] One or more tags. Each tag consists of a tag key and a tag value. The
tag value can be an empty (null) string. You can't have more than one
tag on an Amazon Web Services Payment Cryptography key with the same tag
key. If you specify an existing tag key with a different tag value,
Amazon Web Services Payment Cryptography replaces the current tag value
with the new one.

Don't include personal, confidential or sensitive information in this
field. This field may be displayed in plaintext in CloudTrail logs and
other output.

To use this parameter, you must have
\code{\link[=paymentcryptographycontrolplane_tag_resource]{tag_resource}}
permission in an IAM policy.

Don't include personal, confidential or sensitive information in this
field. This field may be displayed in plaintext in CloudTrail logs and
other output.}
}
\description{
Adds or edits tags on an Amazon Web Services Payment Cryptography key.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_tag_resource/} for full documentation.
}
\keyword{internal}
