% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_untag_resource}
\alias{paymentcryptographycontrolplane_untag_resource}
\title{Deletes a tag from an Amazon Web Services Payment Cryptography key}
\usage{
paymentcryptographycontrolplane_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The \code{KeyARN} of the key whose tags are being removed.}

\item{TagKeys}{[required] One or more tag keys. Don't include the tag values.

If the Amazon Web Services Payment Cryptography key doesn't have the
specified tag key, Amazon Web Services Payment Cryptography doesn't
throw an exception or return a response. To confirm that the operation
succeeded, use the
\code{\link[=paymentcryptographycontrolplane_list_tags_for_resource]{list_tags_for_resource}}
operation.}
}
\description{
Deletes a tag from an Amazon Web Services Payment Cryptography key.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_untag_resource/} for full documentation.
}
\keyword{internal}
