% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_update_alias}
\alias{paymentcryptographycontrolplane_update_alias}
\title{Associates an existing Amazon Web Services Payment Cryptography alias
with a different key}
\usage{
paymentcryptographycontrolplane_update_alias(AliasName, KeyArn = NULL)
}
\arguments{
\item{AliasName}{[required] The alias whose associated key is changing.}

\item{KeyArn}{The \code{KeyARN} for the key that you are updating or removing from the
alias.}
}
\description{
Associates an existing Amazon Web Services Payment Cryptography alias with a different key. Each alias is associated with only one Amazon Web Services Payment Cryptography key at a time, although a key can have multiple aliases. The alias and the Amazon Web Services Payment Cryptography key must be in the same Amazon Web Services account and Amazon Web Services Region

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_update_alias/} for full documentation.
}
\keyword{internal}
