% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographydataplane_operations.R
\name{paymentcryptographydataplane_generate_auth_request_cryptogram}
\alias{paymentcryptographydataplane_generate_auth_request_cryptogram}
\title{Generates an Authorization Request Cryptogram (ARQC) for an EMV chip
payment card authorization}
\usage{
paymentcryptographydataplane_generate_auth_request_cryptogram(
  KeyIdentifier,
  TransactionData,
  MajorKeyDerivationMode,
  SessionKeyDerivationAttributes
)
}
\arguments{
\item{KeyIdentifier}{[required] The \code{keyARN} of the IMK-AC (TR31_E0_EMV_MKEY_APP_CRYPTOGRAMS) that Amazon Web Services Payment Cryptography uses to generate the ARQC.}

\item{TransactionData}{[required] The transaction data that Amazon Web Services Payment Cryptography uses for ARQC generation. The same transaction data is used for ARQC verification by the issuer using \code{\link[=paymentcryptographydataplane_verify_auth_request_cryptogram]{verify_auth_request_cryptogram}}.}

\item{MajorKeyDerivationMode}{[required] The method to use when deriving the major encryption key for ARQC generation within Amazon Web Services Payment Cryptography.}

\item{SessionKeyDerivationAttributes}{[required] The attributes and values to use for deriving a session key for ARQC generation within Amazon Web Services Payment Cryptography.}
}
\description{
Generates an Authorization Request Cryptogram (ARQC) for an EMV chip payment card authorization. For more information, see \href{https://docs.aws.amazon.com/payment-cryptography/latest/userguide/data-operations.generateauthrequestcryptogram.html}{Generate auth request cryptogram} in the \emph{Amazon Web Services Payment Cryptography User Guide}.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographydataplane_generate_auth_request_cryptogram/} for full documentation.
}
\keyword{internal}
