% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographydataplane_operations.R
\name{paymentcryptographydataplane_translate_key_material}
\alias{paymentcryptographydataplane_translate_key_material}
\title{Translates an cryptographic key between different wrapping keys without
importing the key into Amazon Web Services Payment Cryptography}
\usage{
paymentcryptographydataplane_translate_key_material(
  IncomingKeyMaterial,
  OutgoingKeyMaterial,
  KeyCheckValueAlgorithm = NULL
)
}
\arguments{
\item{IncomingKeyMaterial}{[required] Parameter information of the TR31WrappedKeyBlock containing the
transaction key.}

\item{OutgoingKeyMaterial}{[required] Parameter information of the wrapping key used to wrap the transaction
key in the outgoing TR31WrappedKeyBlock.}

\item{KeyCheckValueAlgorithm}{The key check value (KCV) algorithm used for calculating the KCV of the
derived key.}
}
\description{
Translates an cryptographic key between different wrapping keys without importing the key into Amazon Web Services Payment Cryptography.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographydataplane_translate_key_material/} for full documentation.
}
\keyword{internal}
