% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographydataplane_operations.R
\name{paymentcryptographydataplane_translate_pin_data}
\alias{paymentcryptographydataplane_translate_pin_data}
\title{Translates encrypted PIN block from and to ISO 9564 formats 0,1,3,4}
\usage{
paymentcryptographydataplane_translate_pin_data(
  EncryptedPinBlock,
  IncomingDukptAttributes = NULL,
  IncomingKeyIdentifier,
  IncomingTranslationAttributes,
  OutgoingDukptAttributes = NULL,
  OutgoingKeyIdentifier,
  OutgoingTranslationAttributes
)
}
\arguments{
\item{EncryptedPinBlock}{[required] The encrypted PIN block data that Amazon Web Services Payment
Cryptography translates.}

\item{IncomingDukptAttributes}{The attributes and values to use for incoming DUKPT encryption key for
PIN block tranlation.}

\item{IncomingKeyIdentifier}{[required] The \code{keyARN} of the encryption key under which incoming PIN block data
is encrypted. This key type can be PEK or BDK.}

\item{IncomingTranslationAttributes}{[required] The format of the incoming PIN block data for tranlation within Amazon
Web Services Payment Cryptography.}

\item{OutgoingDukptAttributes}{The attributes and values to use for outgoing DUKPT encryption key after
PIN block translation.}

\item{OutgoingKeyIdentifier}{[required] The \code{keyARN} of the encryption key for encrypting outgoing PIN block
data. This key type can be PEK or BDK.}

\item{OutgoingTranslationAttributes}{[required] The format of the outgoing PIN block data after tranlation by Amazon Web
Services Payment Cryptography.}
}
\description{
Translates encrypted PIN block from and to ISO 9564 formats 0,1,3,4. For more information, see \href{https://docs.aws.amazon.com/payment-cryptography/latest/userguide/translate-pin-data.html}{Translate PIN data} in the \emph{Amazon Web Services Payment Cryptography User Guide}.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographydataplane_translate_pin_data/} for full documentation.
}
\keyword{internal}
