% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographydataplane_operations.R
\name{paymentcryptographydataplane_verify_mac}
\alias{paymentcryptographydataplane_verify_mac}
\title{Verifies a Message Authentication Code (MAC)}
\usage{
paymentcryptographydataplane_verify_mac(
  KeyIdentifier,
  MessageData,
  Mac,
  VerificationAttributes,
  MacLength = NULL
)
}
\arguments{
\item{KeyIdentifier}{[required] The \code{keyARN} of the encryption key that Amazon Web Services Payment
Cryptography uses to verify MAC data.}

\item{MessageData}{[required] The data on for which MAC is under verification. This value must be
hexBinary.}

\item{Mac}{[required] The MAC being verified.}

\item{VerificationAttributes}{[required] The attributes and data values to use for MAC verification within Amazon
Web Services Payment Cryptography.}

\item{MacLength}{The length of the MAC.}
}
\description{
Verifies a Message Authentication Code (MAC).

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographydataplane_verify_mac/} for full documentation.
}
\keyword{internal}
