% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_describe_savings_plan_rates}
\alias{savingsplans_describe_savings_plan_rates}
\title{Describes the rates for a specific, existing Savings Plan}
\usage{
savingsplans_describe_savings_plan_rates(
  savingsPlanId,
  filters = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{savingsPlanId}{[required] The ID of the Savings Plan.}

\item{filters}{The filters.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
additional results, make another call with the returned token value.}
}
\description{
Describes the rates for a specific, existing Savings Plan.

See \url{https://www.paws-r-sdk.com/docs/savingsplans_describe_savings_plan_rates/} for full documentation.
}
\keyword{internal}
