% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_describe_savings_plans}
\alias{savingsplans_describe_savings_plans}
\title{Describes the specified Savings Plans}
\usage{
savingsplans_describe_savings_plans(
  savingsPlanArns = NULL,
  savingsPlanIds = NULL,
  nextToken = NULL,
  maxResults = NULL,
  states = NULL,
  filters = NULL
)
}
\arguments{
\item{savingsPlanArns}{The Amazon Resource Names (ARN) of the Savings Plans.}

\item{savingsPlanIds}{The IDs of the Savings Plans.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
additional results, make another call with the returned token value.}

\item{states}{The states.}

\item{filters}{The filters.}
}
\description{
Describes the specified Savings Plans.

See \url{https://paws-r.github.io/docs/savingsplans/describe_savings_plans.html} for full documentation.
}
\keyword{internal}
