% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_describe_savings_plans_offering_rates}
\alias{savingsplans_describe_savings_plans_offering_rates}
\title{Describes the offering rates for Savings Plans you might want to
purchase}
\usage{
savingsplans_describe_savings_plans_offering_rates(
  savingsPlanOfferingIds = NULL,
  savingsPlanPaymentOptions = NULL,
  savingsPlanTypes = NULL,
  products = NULL,
  serviceCodes = NULL,
  usageTypes = NULL,
  operations = NULL,
  filters = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{savingsPlanOfferingIds}{The IDs of the offerings.}

\item{savingsPlanPaymentOptions}{The payment options.}

\item{savingsPlanTypes}{The plan types.}

\item{products}{The Amazon Web Services products.}

\item{serviceCodes}{The services.}

\item{usageTypes}{The usage details of the line item in the billing report.}

\item{operations}{The specific Amazon Web Services operation for the line item in the
billing report.}

\item{filters}{The filters.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
additional results, make another call with the returned token value.}
}
\description{
Describes the offering rates for Savings Plans you might want to purchase.

See \url{https://www.paws-r-sdk.com/docs/savingsplans_describe_savings_plans_offering_rates/} for full documentation.
}
\keyword{internal}
