% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_describe_savings_plans_offerings}
\alias{savingsplans_describe_savings_plans_offerings}
\title{Describes the specified Savings Plans offerings}
\usage{
savingsplans_describe_savings_plans_offerings(
  offeringIds = NULL,
  paymentOptions = NULL,
  productType = NULL,
  planTypes = NULL,
  durations = NULL,
  currencies = NULL,
  descriptions = NULL,
  serviceCodes = NULL,
  usageTypes = NULL,
  operations = NULL,
  filters = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{offeringIds}{The IDs of the offerings.}

\item{paymentOptions}{The payment options.}

\item{productType}{The product type.}

\item{planTypes}{The plan type.}

\item{durations}{The durations, in seconds.}

\item{currencies}{The currencies.}

\item{descriptions}{The descriptions.}

\item{serviceCodes}{The services.}

\item{usageTypes}{The usage details of the line item in the billing report.}

\item{operations}{The specific AWS operation for the line item in the billing report.}

\item{filters}{The filters.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
additional results, make another call with the returned token value.}
}
\description{
Describes the specified Savings Plans offerings.

See \url{https://paws-r.github.io/docs/savingsplans/describe_savings_plans_offerings.html} for full documentation.
}
\keyword{internal}
