% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_list_tags_for_resource}
\alias{savingsplans_list_tags_for_resource}
\title{Lists the tags for the specified resource}
\usage{
savingsplans_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Lists the tags for the specified resource.

See \url{https://www.paws-r-sdk.com/docs/savingsplans_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
