% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingsplans_operations.R
\name{savingsplans_return_savings_plan}
\alias{savingsplans_return_savings_plan}
\title{Returns the specified Savings Plan}
\usage{
savingsplans_return_savings_plan(savingsPlanId, clientToken = NULL)
}
\arguments{
\item{savingsPlanId}{[required] The ID of the Savings Plan.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Returns the specified Savings Plan.

See \url{https://www.paws-r-sdk.com/docs/savingsplans_return_savings_plan/} for full documentation.
}
\keyword{internal}
