% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_service.R
\name{billing}
\alias{billing}
\title{AWS Billing}
\usage{
billing(config = list(), credentials = list(), endpoint = NULL, region = NULL)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
\item{\strong{use_dual_stack}: Set this to \code{true} to use the dualstack (IPv4 and IPv6) endpoint for a service, where available, falling back to the regular endpoint when it isn't. Defaults to the \code{AWS_USE_DUALSTACK_ENDPOINT} environment variable when unset.}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
You can use the Billing API to programatically list the billing views available to you for a given time period. A billing view represents a set of billing data.

The Billing API provides the following endpoint:

\verb{https://billing.us-east-1.api.aws}
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- billing(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string",
    use_dual_stack = "logical"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=billing_associate_source_views]{associate_source_views} \tab Associates one or more source billing views with an existing billing view\cr
\link[=billing_create_billing_view]{create_billing_view} \tab Creates a billing view with the specified billing view attributes\cr
\link[=billing_delete_billing_view]{delete_billing_view} \tab Deletes the specified billing view\cr
\link[=billing_disassociate_source_views]{disassociate_source_views} \tab Removes the association between one or more source billing views and an existing billing view\cr
\link[=billing_get_billing_preferences]{get_billing_preferences} \tab Retrieves billing preferences for the specified feature\cr
\link[=billing_get_billing_view]{get_billing_view} \tab Returns the metadata associated to the specified billing view ARN\cr
\link[=billing_get_credit_allocation_history]{get_credit_allocation_history} \tab Returns the per-billing-month allocation history for credits applied to an Amazon Web Services account's bills\cr
\link[=billing_get_credits]{get_credits} \tab Returns the list of Amazon Web Services account credits for the specified account\cr
\link[=billing_get_enterprise_support_charge_summary]{get_enterprise_support_charge_summary} \tab Returns a summary of Enterprise Support data aggregated across all accounts in the Enterprise Support profile\cr
\link[=billing_get_enterprise_support_contract_details]{get_enterprise_support_contract_details} \tab Returns Enterprise Support contract details\cr
\link[=billing_get_resource_policy]{get_resource_policy} \tab Returns the resource-based policy document attached to the resource in JSON format\cr
\link[=billing_list_billing_views]{list_billing_views} \tab Lists the billing views available for a given time period\cr
\link[=billing_list_billing_view_segments]{list_billing_view_segments} \tab Lists the segments of a billing view over a given time period\cr
\link[=billing_list_business_support_account_charges]{list_business_support_account_charges} \tab Returns Business Support charges broken down at the linked account level for a given billing month\cr
\link[=billing_list_business_support_subscription_history]{list_business_support_subscription_history} \tab Returns the history of Business Support subscription contracts across accounts\cr
\link[=billing_list_enterprise_support_linked_account_charges]{list_enterprise_support_linked_account_charges} \tab Returns Support-eligible spend broken down at linked account level\cr
\link[=billing_list_source_views_for_billing_view]{list_source_views_for_billing_view} \tab Lists the source views (managed Amazon Web Services billing views) associated with the billing view\cr
\link[=billing_list_tags_for_resource]{list_tags_for_resource} \tab Lists tags associated with the billing view resource\cr
\link[=billing_redeem_credits]{redeem_credits} \tab Redeems an Amazon Web Services promotional credit code on behalf of the calling account\cr
\link[=billing_tag_resource]{tag_resource} \tab An API operation for adding one or more tags (key-value pairs) to a resource\cr
\link[=billing_untag_resource]{untag_resource} \tab Removes one or more tags from a resource\cr
\link[=billing_update_billing_preferences]{update_billing_preferences} \tab Updates billing preferences for the specified feature\cr
\link[=billing_update_billing_view]{update_billing_view} \tab An API to update the attributes of the billing view
}
}

\examples{
\dontrun{
svc <- billing()
svc$associate_source_views(
  Foo = 123
)
}

}
