% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_batch_associate_resources_to_custom_line_item}
\alias{billingconductor_batch_associate_resources_to_custom_line_item}
\title{Associates a batch of resources to a percentage custom line item}
\usage{
billingconductor_batch_associate_resources_to_custom_line_item(
  TargetArn,
  ResourceArns,
  BillingPeriodRange = NULL
)
}
\arguments{
\item{TargetArn}{[required] A percentage custom line item ARN to associate the resources to.}

\item{ResourceArns}{[required] A list containing the ARNs of the resources to be associated.}

\item{BillingPeriodRange}{}
}
\description{
Associates a batch of resources to a percentage custom line item.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_batch_associate_resources_to_custom_line_item/} for full documentation.
}
\keyword{internal}
