% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billingconductor_operations.R
\name{billingconductor_update_pricing_rule}
\alias{billingconductor_update_pricing_rule}
\title{Updates an existing pricing rule}
\usage{
billingconductor_update_pricing_rule(
  Arn,
  Name = NULL,
  Description = NULL,
  Type = NULL,
  ModifierPercentage = NULL,
  Tiering = NULL
)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the pricing rule to update.}

\item{Name}{The new name of the pricing rule. The name must be unique to each
pricing rule.}

\item{Description}{The new description for the pricing rule.}

\item{Type}{The new pricing rule type.}

\item{ModifierPercentage}{The new modifier to show pricing plan rates as a percentage. Your entry
will be rounded to the nearest 2 decimal places.}

\item{Tiering}{The set of tiering configurations for the pricing rule.}
}
\description{
Updates an existing pricing rule.

See \url{https://www.paws-r-sdk.com/docs/billingconductor_update_pricing_rule/} for full documentation.
}
\keyword{internal}
