% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_create_budget_action}
\alias{budgets_create_budget_action}
\title{Creates a budget action}
\usage{
budgets_create_budget_action(
  AccountId,
  BudgetName,
  NotificationType,
  ActionType,
  ActionThreshold,
  Definition,
  ExecutionRoleArn,
  ApprovalModel,
  Subscribers,
  ResourceTags = NULL
)
}
\arguments{
\item{AccountId}{[required]}

\item{BudgetName}{[required]}

\item{NotificationType}{[required]}

\item{ActionType}{[required] The type of action. This defines the type of tasks that can be carried
out by this action. This field also determines the format for
definition.}

\item{ActionThreshold}{[required]}

\item{Definition}{[required]}

\item{ExecutionRoleArn}{[required] The role passed for action execution and reversion. Roles and actions
must be in the same account.}

\item{ApprovalModel}{[required] This specifies if the action needs manual or automatic approval.}

\item{Subscribers}{[required]}

\item{ResourceTags}{An optional list of tags to associate with the specified budget action.
Each tag consists of a key and a value, and each key must be unique for
the resource.}
}
\description{
Creates a budget action.

See \url{https://www.paws-r-sdk.com/docs/budgets_create_budget_action/} for full documentation.
}
\keyword{internal}
