% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_delete_budget}
\alias{budgets_delete_budget}
\title{Deletes a budget}
\usage{
budgets_delete_budget(AccountId, BudgetName)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget that you want to
delete.}

\item{BudgetName}{[required] The name of the budget that you want to delete.}
}
\value{
An empty list.
}
\description{
Deletes a budget. You can delete your budget at any time.

Deleting a budget also deletes the notifications and subscribers that
are associated with that budget.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_budget(
  AccountId = "string",
  BudgetName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
