% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_delete_subscriber}
\alias{budgets_delete_subscriber}
\title{Deletes a subscriber}
\usage{
budgets_delete_subscriber(AccountId, BudgetName, Notification, Subscriber)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget whose subscriber you
want to delete.}

\item{BudgetName}{[required] The name of the budget whose subscriber you want to delete.}

\item{Notification}{[required] The notification whose subscriber you want to delete.}

\item{Subscriber}{[required] The subscriber that you want to delete.}
}
\description{
Deletes a subscriber.

See \url{https://www.paws-r-sdk.com/docs/budgets_delete_subscriber/} for full documentation.
}
\keyword{internal}
