% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget}
\alias{budgets_describe_budget}
\title{Describes a budget}
\usage{
budgets_describe_budget(AccountId, BudgetName)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget that you want a
description of.}

\item{BudgetName}{[required] The name of the budget that you want a description of.}
}
\description{
Describes a budget.

See \url{https://paws-r.github.io/docs/budgets/describe_budget.html} for full documentation.
}
\keyword{internal}
