% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget_action}
\alias{budgets_describe_budget_action}
\title{Describes a budget action detail}
\usage{
budgets_describe_budget_action(AccountId, BudgetName, ActionId)
}
\arguments{
\item{AccountId}{[required]}

\item{BudgetName}{[required]}

\item{ActionId}{[required] A system-generated universally unique identifier (UUID) for the action.}
}
\description{
Describes a budget action detail.
}
\section{Request syntax}{
\preformatted{svc$describe_budget_action(
  AccountId = "string",
  BudgetName = "string",
  ActionId = "string"
)
}
}

\keyword{internal}
