% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget_action}
\alias{budgets_describe_budget_action}
\title{Describes a budget action detail}
\usage{
budgets_describe_budget_action(AccountId, BudgetName, ActionId)
}
\arguments{
\item{AccountId}{[required]}

\item{BudgetName}{[required]}

\item{ActionId}{[required] A system-generated universally unique identifier (UUID) for the action.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AccountId = "string",
  BudgetName = "string",
  Action = list(
    ActionId = "string",
    BudgetName = "string",
    NotificationType = "ACTUAL"|"FORECASTED",
    ActionType = "APPLY_IAM_POLICY"|"APPLY_SCP_POLICY"|"RUN_SSM_DOCUMENTS",
    ActionThreshold = list(
      ActionThresholdValue = 123.0,
      ActionThresholdType = "PERCENTAGE"|"ABSOLUTE_VALUE"
    ),
    Definition = list(
      IamActionDefinition = list(
        PolicyArn = "string",
        Roles = list(
          "string"
        ),
        Groups = list(
          "string"
        ),
        Users = list(
          "string"
        )
      ),
      ScpActionDefinition = list(
        PolicyId = "string",
        TargetIds = list(
          "string"
        )
      ),
      SsmActionDefinition = list(
        ActionSubType = "STOP_EC2_INSTANCES"|"STOP_RDS_INSTANCES",
        Region = "string",
        InstanceIds = list(
          "string"
        )
      )
    ),
    ExecutionRoleArn = "string",
    ApprovalModel = "AUTOMATIC"|"MANUAL",
    Status = "STANDBY"|"PENDING"|"EXECUTION_IN_PROGRESS"|"EXECUTION_SUCCESS"|"EXECUTION_FAILURE"|"REVERSE_IN_PROGRESS"|"REVERSE_SUCCESS"|"REVERSE_FAILURE"|"RESET_IN_PROGRESS"|"RESET_FAILURE",
    Subscribers = list(
      list(
        SubscriptionType = "SNS"|"EMAIL",
        Address = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a budget action detail.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_budget_action(
  AccountId = "string",
  BudgetName = "string",
  ActionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
