% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget_action_histories}
\alias{budgets_describe_budget_action_histories}
\title{Describes a budget action history detail}
\usage{
budgets_describe_budget_action_histories(AccountId, BudgetName,
  ActionId, TimePeriod, MaxResults, NextToken)
}
\arguments{
\item{AccountId}{[required]}

\item{BudgetName}{[required]}

\item{ActionId}{[required] A system-generated universally unique identifier (UUID) for the action.}

\item{TimePeriod}{}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Describes a budget action history detail.
}
\section{Request syntax}{
\preformatted{svc$describe_budget_action_histories(
  AccountId = "string",
  BudgetName = "string",
  ActionId = "string",
  TimePeriod = list(
    Start = as.POSIXct(
      "2015-01-01"
    ),
    End = as.POSIXct(
      "2015-01-01"
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
