% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget_notifications_for_account}
\alias{budgets_describe_budget_notifications_for_account}
\title{Lists the budget names and notifications that are associated with an
account}
\usage{
budgets_describe_budget_notifications_for_account(
  AccountId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AccountId}{[required]}

\item{MaxResults}{An integer that shows how many budget name entries a paginated response
contains.}

\item{NextToken}{}
}
\description{
Lists the budget names and notifications that are associated with an account.

See \url{https://paws-r.github.io/docs/budgets/describe_budget_notifications_for_account.html} for full documentation.
}
\keyword{internal}
