% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget_performance_history}
\alias{budgets_describe_budget_performance_history}
\title{Describes the history for DAILY, MONTHLY, and QUARTERLY budgets}
\usage{
budgets_describe_budget_performance_history(
  AccountId,
  BudgetName,
  TimePeriod = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AccountId}{[required]}

\item{BudgetName}{[required]}

\item{TimePeriod}{Retrieves how often the budget went into an \code{ALARM} state for the
specified time period.}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Describes the history for \code{DAILY}, \code{MONTHLY}, and \code{QUARTERLY} budgets. Budget history isn't available for \code{ANNUAL} budgets.

See \url{https://www.paws-r-sdk.com/docs/budgets_describe_budget_performance_history/} for full documentation.
}
\keyword{internal}
