% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget_performance_history}
\alias{budgets_describe_budget_performance_history}
\title{Describes the history for DAILY, MONTHLY, and QUARTERLY budgets}
\usage{
budgets_describe_budget_performance_history(AccountId, BudgetName,
  TimePeriod, MaxResults, NextToken)
}
\arguments{
\item{AccountId}{[required]}

\item{BudgetName}{[required]}

\item{TimePeriod}{Retrieves how often the budget went into an \code{ALARM} state for the
specified time period.}

\item{MaxResults}{}

\item{NextToken}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BudgetPerformanceHistory = list(
    BudgetName = "string",
    BudgetType = "USAGE"|"COST"|"RI_UTILIZATION"|"RI_COVERAGE"|"SAVINGS_PLANS_UTILIZATION"|"SAVINGS_PLANS_COVERAGE",
    CostFilters = list(
      list(
        "string"
      )
    ),
    CostTypes = list(
      IncludeTax = TRUE|FALSE,
      IncludeSubscription = TRUE|FALSE,
      UseBlended = TRUE|FALSE,
      IncludeRefund = TRUE|FALSE,
      IncludeCredit = TRUE|FALSE,
      IncludeUpfront = TRUE|FALSE,
      IncludeRecurring = TRUE|FALSE,
      IncludeOtherSubscription = TRUE|FALSE,
      IncludeSupport = TRUE|FALSE,
      IncludeDiscount = TRUE|FALSE,
      UseAmortized = TRUE|FALSE
    ),
    TimeUnit = "DAILY"|"MONTHLY"|"QUARTERLY"|"ANNUALLY",
    BudgetedAndActualAmountsList = list(
      list(
        BudgetedAmount = list(
          Amount = "string",
          Unit = "string"
        ),
        ActualAmount = list(
          Amount = "string",
          Unit = "string"
        ),
        TimePeriod = list(
          Start = as.POSIXct(
            "2015-01-01"
          ),
          End = as.POSIXct(
            "2015-01-01"
          )
        )
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Describes the history for \code{DAILY}, \code{MONTHLY}, and \code{QUARTERLY} budgets.
Budget history isn't available for \code{ANNUAL} budgets.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_budget_performance_history(
  AccountId = "string",
  BudgetName = "string",
  TimePeriod = list(
    Start = as.POSIXct(
      "2015-01-01"
    ),
    End = as.POSIXct(
      "2015-01-01"
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
