% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budgets}
\alias{budgets_describe_budgets}
\title{Lists the budgets that are associated with an account}
\usage{
budgets_describe_budgets(AccountId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budgets that you want to
describe.}

\item{MaxResults}{An integer that represents how many budgets a paginated response
contains. The default is 100.}

\item{NextToken}{The pagination token that you include in your request to indicate the
next set of results that you want to retrieve.}
}
\description{
Lists the budgets that are associated with an account.

See \url{https://www.paws-r-sdk.com/docs/budgets_describe_budgets/} for full documentation.
}
\keyword{internal}
