% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_execute_budget_action}
\alias{budgets_execute_budget_action}
\title{Executes a budget action}
\usage{
budgets_execute_budget_action(AccountId, BudgetName, ActionId, ExecutionType)
}
\arguments{
\item{AccountId}{[required]}

\item{BudgetName}{[required]}

\item{ActionId}{[required] A system-generated universally unique identifier (UUID) for the action.}

\item{ExecutionType}{[required] The type of execution.}
}
\description{
Executes a budget action.

See \url{https://www.paws-r-sdk.com/docs/budgets_execute_budget_action/} for full documentation.
}
\keyword{internal}
