% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_execute_budget_action}
\alias{budgets_execute_budget_action}
\title{Executes a budget action}
\usage{
budgets_execute_budget_action(AccountId, BudgetName, ActionId,
  ExecutionType)
}
\arguments{
\item{AccountId}{[required]}

\item{BudgetName}{[required]}

\item{ActionId}{[required] A system-generated universally unique identifier (UUID) for the action.}

\item{ExecutionType}{[required] The type of execution.}
}
\value{
A list with the following syntax:\preformatted{list(
  AccountId = "string",
  BudgetName = "string",
  ActionId = "string",
  ExecutionType = "APPROVE_BUDGET_ACTION"|"RETRY_BUDGET_ACTION"|"REVERSE_BUDGET_ACTION"|"RESET_BUDGET_ACTION"
)
}
}
\description{
Executes a budget action.
}
\section{Request syntax}{
\preformatted{svc$execute_budget_action(
  AccountId = "string",
  BudgetName = "string",
  ActionId = "string",
  ExecutionType = "APPROVE_BUDGET_ACTION"|"RETRY_BUDGET_ACTION"|"REVERSE_BUDGET_ACTION"|"RESET_BUDGET_ACTION"
)
}
}

\keyword{internal}
