% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_update_budget_action}
\alias{budgets_update_budget_action}
\title{Updates a budget action}
\usage{
budgets_update_budget_action(
  AccountId,
  BudgetName,
  ActionId,
  NotificationType = NULL,
  ActionThreshold = NULL,
  Definition = NULL,
  ExecutionRoleArn = NULL,
  ApprovalModel = NULL,
  Subscribers = NULL
)
}
\arguments{
\item{AccountId}{[required]}

\item{BudgetName}{[required]}

\item{ActionId}{[required] A system-generated universally unique identifier (UUID) for the action.}

\item{NotificationType}{}

\item{ActionThreshold}{}

\item{Definition}{}

\item{ExecutionRoleArn}{The role passed for action execution and reversion. Roles and actions
must be in the same account.}

\item{ApprovalModel}{This specifies if the action needs manual or automatic approval.}

\item{Subscribers}{}
}
\description{
Updates a budget action.

See \url{https://paws-r.github.io/docs/budgets/update_budget_action.html} for full documentation.
}
\keyword{internal}
