% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_update_notification}
\alias{budgets_update_notification}
\title{Updates a notification}
\usage{
budgets_update_notification(
  AccountId,
  BudgetName,
  OldNotification,
  NewNotification
)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget whose notification
you want to update.}

\item{BudgetName}{[required] The name of the budget whose notification you want to update.}

\item{OldNotification}{[required] The previous notification that is associated with a budget.}

\item{NewNotification}{[required] The updated notification to be associated with a budget.}
}
\description{
Updates a notification.

See \url{https://www.paws-r-sdk.com/docs/budgets_update_notification/} for full documentation.
}
\keyword{internal}
