% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_describe_report_definitions}
\alias{costandusagereportservice_describe_report_definitions}
\title{Lists the AWS Cost and Usage reports available to this account}
\usage{
costandusagereportservice_describe_report_definitions(MaxResults,
  NextToken)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ReportDefinitions = list(
    list(
      ReportName = "string",
      TimeUnit = "HOURLY"|"DAILY"|"MONTHLY",
      Format = "textORcsv"|"Parquet",
      Compression = "ZIP"|"GZIP"|"Parquet",
      AdditionalSchemaElements = list(
        "RESOURCES"
      ),
      S3Bucket = "string",
      S3Prefix = "string",
      S3Region = "af-south-1"|"ap-east-1"|"ap-south-1"|"ap-southeast-1"|"ap-southeast-2"|"ap-northeast-1"|"ap-northeast-2"|"ap-northeast-3"|"ca-central-1"|"eu-central-1"|"eu-west-1"|"eu-west-2"|"eu-west-3"|"eu-north-1"|"eu-south-1"|"me-south-1"|"sa-east-1"|"us-east-1"|"us-east-2"|"us-west-1"|"us-west-2"|"cn-north-1"|"cn-northwest-1",
      AdditionalArtifacts = list(
        "REDSHIFT"|"QUICKSIGHT"|"ATHENA"
      ),
      RefreshClosedReports = TRUE|FALSE,
      ReportVersioning = "CREATE_NEW_REPORT"|"OVERWRITE_REPORT"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the AWS Cost and Usage reports available to this account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_report_definitions(
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example lists the AWS Cost and Usage reports for the
# account.
svc$describe_report_definitions(
  MaxResults = 5L
)
}

}
\keyword{internal}
