% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_describe_report_definitions}
\alias{costandusagereportservice_describe_report_definitions}
\title{Lists the Amazon Web Services Cost and Usage Report available to this
account}
\usage{
costandusagereportservice_describe_report_definitions(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Lists the Amazon Web Services Cost and Usage Report available to this account.

See \url{https://www.paws-r-sdk.com/docs/costandusagereportservice_describe_report_definitions/} for full documentation.
}
\keyword{internal}
