% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_modify_report_definition}
\alias{costandusagereportservice_modify_report_definition}
\title{Allows you to programatically update your report preferences}
\usage{
costandusagereportservice_modify_report_definition(
  ReportName,
  ReportDefinition
)
}
\arguments{
\item{ReportName}{[required]}

\item{ReportDefinition}{[required]}
}
\description{
Allows you to programatically update your report preferences.

See \url{https://paws-r.github.io/docs/costandusagereportservice/modify_report_definition.html} for full documentation.
}
\keyword{internal}
