% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costandusagereportservice_operations.R
\name{costandusagereportservice_modify_report_definition}
\alias{costandusagereportservice_modify_report_definition}
\title{Allows you to programmatically update your report preferences}
\usage{
costandusagereportservice_modify_report_definition(
  ReportName,
  ReportDefinition
)
}
\arguments{
\item{ReportName}{[required]}

\item{ReportDefinition}{[required]}
}
\description{
Allows you to programmatically update your report preferences.

See \url{https://www.paws-r-sdk.com/docs/costandusagereportservice_modify_report_definition/} for full documentation.
}
\keyword{internal}
