% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_delete_anomaly_monitor}
\alias{costexplorer_delete_anomaly_monitor}
\title{Deletes a cost anomaly monitor}
\usage{
costexplorer_delete_anomaly_monitor(MonitorArn)
}
\arguments{
\item{MonitorArn}{[required] The unique identifier of the cost anomaly monitor that you want to
delete.}
}
\description{
Deletes a cost anomaly monitor.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_delete_anomaly_monitor/} for full documentation.
}
\keyword{internal}
