% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_delete_anomaly_subscription}
\alias{costexplorer_delete_anomaly_subscription}
\title{Deletes a cost anomaly subscription}
\usage{
costexplorer_delete_anomaly_subscription(SubscriptionArn)
}
\arguments{
\item{SubscriptionArn}{[required] The unique identifier of the cost anomaly subscription that you want to
delete.}
}
\value{
An empty list.
}
\description{
Deletes a cost anomaly subscription.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_anomaly_subscription(
  SubscriptionArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
