% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_describe_cost_category_definition}
\alias{costexplorer_describe_cost_category_definition}
\title{Returns the name, Amazon Resource Name (ARN), rules, definition, and
effective dates of a cost category that's defined in the account}
\usage{
costexplorer_describe_cost_category_definition(
  CostCategoryArn,
  EffectiveOn = NULL
)
}
\arguments{
\item{CostCategoryArn}{[required] The unique identifier for your cost category.}

\item{EffectiveOn}{The date when the cost category was effective.}
}
\description{
Returns the name, Amazon Resource Name (ARN), rules, definition, and effective dates of a cost category that's defined in the account.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_describe_cost_category_definition/} for full documentation.
}
\keyword{internal}
