% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_get_anomalies}
\alias{costexplorer_get_anomalies}
\title{Retrieves all of the cost anomalies detected on your account, during the
time period specified by the DateInterval object}
\usage{
costexplorer_get_anomalies(MonitorArn, DateInterval, Feedback,
  TotalImpact, NextPageToken, MaxResults)
}
\arguments{
\item{MonitorArn}{Retrieves all of the cost anomalies detected for a specific cost anomaly
monitor Amazon Resource Name (ARN).}

\item{DateInterval}{[required] Assigns the start and end dates for retrieving cost anomalies. The
returned anomaly object will have an \code{AnomalyEndDate} in the specified
time range.}

\item{Feedback}{Filters anomaly results by the feedback field on the anomaly object.}

\item{TotalImpact}{Filters anomaly results by the total impact field on the anomaly object.
For example, you can filter anomalies \verb{GREATER_THAN 200.00} to retrieve
anomalies, with an estimated dollar impact greater than 200.}

\item{NextPageToken}{The token to retrieve the next set of results. AWS provides the token
when the response from a previous call has more results than the maximum
page size.}

\item{MaxResults}{The number of entries a paginated response contains.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Anomalies = list(
    list(
      AnomalyId = "string",
      AnomalyStartDate = "string",
      AnomalyEndDate = "string",
      DimensionValue = "string",
      RootCauses = list(
        list(
          Service = "string",
          Region = "string",
          LinkedAccount = "string",
          UsageType = "string"
        )
      ),
      AnomalyScore = list(
        MaxScore = 123.0,
        CurrentScore = 123.0
      ),
      Impact = list(
        MaxImpact = 123.0,
        TotalImpact = 123.0
      ),
      MonitorArn = "string",
      Feedback = "YES"|"NO"|"PLANNED_ACTIVITY"
    )
  ),
  NextPageToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves all of the cost anomalies detected on your account, during the
time period specified by the \code{DateInterval} object.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_anomalies(
  MonitorArn = "string",
  DateInterval = list(
    StartDate = "string",
    EndDate = "string"
  ),
  Feedback = "YES"|"NO"|"PLANNED_ACTIVITY",
  TotalImpact = list(
    NumericOperator = "EQUAL"|"GREATER_THAN_OR_EQUAL"|"LESS_THAN_OR_EQUAL"|"GREATER_THAN"|"LESS_THAN"|"BETWEEN",
    StartValue = 123.0,
    EndValue = 123.0
  ),
  NextPageToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
