% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_get_anomaly_monitors}
\alias{costexplorer_get_anomaly_monitors}
\title{Retrieves the cost anomaly monitor definitions for your account}
\usage{
costexplorer_get_anomaly_monitors(
  MonitorArnList = NULL,
  NextPageToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{MonitorArnList}{A list of cost anomaly monitor ARNs.}

\item{NextPageToken}{The token to retrieve the next set of results. Amazon Web Services
provides the token when the response from a previous call has more
results than the maximum page size.}

\item{MaxResults}{The number of entries that a paginated response contains.}
}
\description{
Retrieves the cost anomaly monitor definitions for your account. You can filter using a list of cost anomaly monitor Amazon Resource Names (ARNs).

See \url{https://paws-r.github.io/docs/costexplorer/get_anomaly_monitors.html} for full documentation.
}
\keyword{internal}
