% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_get_anomaly_subscriptions}
\alias{costexplorer_get_anomaly_subscriptions}
\title{Retrieves the cost anomaly subscription objects for your account}
\usage{
costexplorer_get_anomaly_subscriptions(
  SubscriptionArnList = NULL,
  MonitorArn = NULL,
  NextPageToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{SubscriptionArnList}{A list of cost anomaly subscription ARNs.}

\item{MonitorArn}{Cost anomaly monitor ARNs.}

\item{NextPageToken}{The token to retrieve the next set of results. Amazon Web Services
provides the token when the response from a previous call has more
results than the maximum page size.}

\item{MaxResults}{The number of entries a paginated response contains.}
}
\description{
Retrieves the cost anomaly subscription objects for your account. You can filter using a list of cost anomaly monitor Amazon Resource Names (ARNs).

See \url{https://www.paws-r-sdk.com/docs/costexplorer_get_anomaly_subscriptions/} for full documentation.
}
\keyword{internal}
