% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_get_cost_and_usage_comparisons}
\alias{costexplorer_get_cost_and_usage_comparisons}
\title{Retrieves cost and usage comparisons for your account between two
periods within the last 13 months}
\usage{
costexplorer_get_cost_and_usage_comparisons(
  BillingViewArn = NULL,
  BaselineTimePeriod,
  ComparisonTimePeriod,
  MetricForComparison,
  Filter = NULL,
  GroupBy = NULL,
  MaxResults = NULL,
  NextPageToken = NULL
)
}
\arguments{
\item{BillingViewArn}{The Amazon Resource Name (ARN) that uniquely identifies a specific
billing view. The ARN is used to specify which particular billing view
you want to interact with or retrieve information from when making API
calls related to Amazon Web Services Billing and Cost Management
features. The BillingViewArn can be retrieved by calling the
ListBillingViews API.}

\item{BaselineTimePeriod}{[required] The reference time period for comparison. This time period serves as the
baseline against which other cost and usage data will be compared. The
interval must start and end on the first day of a month, with a duration
of exactly one month.}

\item{ComparisonTimePeriod}{[required] The comparison time period for analysis. This time period's cost and
usage data will be compared against the baseline time period. The
interval must start and end on the first day of a month, with a duration
of exactly one month.}

\item{MetricForComparison}{[required] The cost and usage metric to compare. Valid values are \code{AmortizedCost},
\code{BlendedCost}, \code{NetAmortizedCost}, \code{NetUnblendedCost},
\code{NormalizedUsageAmount}, \code{UnblendedCost}, and \code{UsageQuantity}.}

\item{Filter}{}

\item{GroupBy}{You can group results using the attributes \code{DIMENSION}, \code{TAG}, and
\code{COST_CATEGORY}.}

\item{MaxResults}{The maximum number of results that are returned for the request.}

\item{NextPageToken}{The token to retrieve the next set of paginated results.}
}
\description{
Retrieves cost and usage comparisons for your account between two periods within the last 13 months. If you have enabled multi-year data at monthly granularity, you can go back up to 38 months.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_get_cost_and_usage_comparisons/} for full documentation.
}
\keyword{internal}
