% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costexplorer_operations.R
\name{costexplorer_get_rightsizing_recommendation}
\alias{costexplorer_get_rightsizing_recommendation}
\title{Creates recommendations that help you save cost by identifying idle and
underutilized Amazon EC2 instances}
\usage{
costexplorer_get_rightsizing_recommendation(
  Filter = NULL,
  Configuration = NULL,
  Service,
  PageSize = NULL,
  NextPageToken = NULL
)
}
\arguments{
\item{Filter}{}

\item{Configuration}{You can use Configuration to customize recommendations across two
attributes. You can choose to view recommendations for instances within
the same instance families or across different instance families. You
can also choose to view your estimated savings that are associated with
recommendations with consideration of existing Savings Plans or RI
benefits, or neither.}

\item{Service}{[required] The specific service that you want recommendations for. The only valid
value for
\code{\link[=costexplorer_get_rightsizing_recommendation]{get_rightsizing_recommendation}}
is "\code{AmazonEC2}".}

\item{PageSize}{The number of recommendations that you want returned in a single
response object.}

\item{NextPageToken}{The pagination token that indicates the next set of results that you
want to retrieve.}
}
\description{
Creates recommendations that help you save cost by identifying idle and underutilized Amazon EC2 instances.

See \url{https://www.paws-r-sdk.com/docs/costexplorer_get_rightsizing_recommendation/} for full documentation.
}
\keyword{internal}
