% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_list_tags_for_resource}
\alias{marketplacecatalog_list_tags_for_resource}
\title{Lists all tags that have been added to a resource (either an entity or
change set)}
\usage{
marketplacecatalog_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] Required. The Amazon Resource Name (ARN) associated with the resource
you want to list tags on.}
}
\description{
Lists all tags that have been added to a resource (either an \href{https://docs.aws.amazon.com/marketplace-catalog/latest/api-reference/welcome.html#catalog-api-entities}{entity} or \href{https://docs.aws.amazon.com/marketplace-catalog/latest/api-reference/welcome.html#working-with-change-sets}{change set}).

See \url{https://paws-r.github.io/docs/marketplacecatalog/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
