% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_tag_resource}
\alias{marketplacecatalog_tag_resource}
\title{Tags a resource (either an entity or change set)}
\usage{
marketplacecatalog_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] Required. The Amazon Resource Name (ARN) associated with the resource
you want to tag.}

\item{Tags}{[required] Required. A list of objects specifying each key name and value. Number
of objects allowed: 1-50.}
}
\description{
Tags a resource (either an \href{https://docs.aws.amazon.com/marketplace-catalog/latest/api-reference/welcome.html#catalog-api-entities}{entity} or \href{https://docs.aws.amazon.com/marketplace-catalog/latest/api-reference/welcome.html#working-with-change-sets}{change set}).

See \url{https://paws-r.github.io/docs/marketplacecatalog/tag_resource.html} for full documentation.
}
\keyword{internal}
