% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacecatalog_operations.R
\name{marketplacecatalog_untag_resource}
\alias{marketplacecatalog_untag_resource}
\title{Removes a tag or list of tags from a resource (either an entity or
change set)}
\usage{
marketplacecatalog_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] Required. The Amazon Resource Name (ARN) associated with the resource
you want to remove the tag from.}

\item{TagKeys}{[required] Required. A list of key names of tags to be removed. Number of strings
allowed: 0-256.}
}
\description{
Removes a tag or list of tags from a resource (either an \href{https://docs.aws.amazon.com/marketplace-catalog/latest/api-reference/welcome.html#catalog-api-entities}{entity} or \href{https://docs.aws.amazon.com/marketplace-catalog/latest/api-reference/welcome.html#working-with-change-sets}{change set}).

See \url{https://paws-r.github.io/docs/marketplacecatalog/untag_resource.html} for full documentation.
}
\keyword{internal}
