% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacemetering_service.R
\name{marketplacemetering}
\alias{marketplacemetering}
\title{AWSMarketplace Metering}
\usage{
marketplacemetering(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Amazon Web Services Marketplace Metering Service

This reference provides descriptions of the low-level Marketplace
Metering Service API.

Amazon Web Services Marketplace sellers can use this API to submit usage
data for custom usage dimensions.

For information about the permissions that you need to use this API, see
\href{https://docs.aws.amazon.com/marketplace/latest/userguide/iam-user-policy-for-aws-marketplace-actions.html}{Amazon Web Services Marketplace metering and entitlement API permissions}
in the \emph{Amazon Web Services Marketplace Seller Guide.}

\strong{Submitting metering records}

\emph{MeterUsage}
\itemize{
\item Submits the metering record for an Amazon Web Services Marketplace
product.
\item Called from: Amazon Elastic Compute Cloud (Amazon EC2) instance or a
container running on either Amazon Elastic Kubernetes Service
(Amazon EKS) or Amazon Elastic Container Service (Amazon ECS)
\item Supported product types: Amazon Machine Images (AMIs) and containers
\item Vendor-metered tagging: Supported allocation tagging
}

\emph{BatchMeterUsage}
\itemize{
\item Submits the metering record for a set of customers.
\code{\link[=marketplacemetering_batch_meter_usage]{batch_meter_usage}} API
calls are captured by CloudTrail. You can use CloudTrail to verify
that the software as a subscription (SaaS) metering records that you
sent are accurate by searching for records with the \code{eventName} of
\code{\link[=marketplacemetering_batch_meter_usage]{batch_meter_usage}}. You
can also use CloudTrail to audit records over time. For more
information, see the \href{https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-concepts.html}{CloudTrail User Guide}.
\item Called from: SaaS applications
\item Supported product type: SaaS
\item Vendor-metered tagging: Supports allocation tagging
}

\strong{Accepting new customers}

\emph{ResolveCustomer}
\itemize{
\item Resolves the registration token that the buyer submits through the
browser during the registration process. Obtains a
\code{CustomerIdentifier} along with the \code{CustomerAWSAccountId} and
\code{ProductCode}.
\item Called from: SaaS application during the registration process
\item Supported product type: SaaS
\item Vendor-metered tagging: Not applicable
}

\strong{Entitlement and metering for paid container products}

\emph{RegisteredUsage}
\itemize{
\item Provides software entitlement and metering. Paid container software
products sold through Amazon Web Services Marketplace must integrate
with the Marketplace Metering Service and call the
\code{\link[=marketplacemetering_register_usage]{register_usage}} operation.
Free and Bring Your Own License model (BYOL) products for Amazon ECS
or Amazon EKS aren't required to call
\code{\link[=marketplacemetering_register_usage]{register_usage}}. However, you
can do so if you want to receive usage data in your seller reports.
For more information about using the
\code{\link[=marketplacemetering_register_usage]{register_usage}} operation,
see \href{https://docs.aws.amazon.com/marketplace/latest/userguide/container-based-products.html}{Container-based products}.
\item Called from: Paid container software products
\item Supported product type: Containers
\item Vendor-metered tagging: Not applicable
}

\strong{Entitlement custom metering for container products}
\itemize{
\item MeterUsage API is available in GovCloud Regions but only supports
AMI FCP products in GovCloud Regions. Flexible Consumption Pricing
(FCP) Container products aren’t supported in GovCloud Regions:
us-gov-west-1 and us-gov-east-1. For more information, see
\href{https://docs.aws.amazon.com/marketplace/latest/userguide/container-based-products.html}{Container-based products}.
\item Custom metering for container products are called using the
MeterUsage API. The API is used for FCP AMI and FCP Container
product metering.
}

\strong{Custom metering for Amazon EKS is available in 17 Amazon Web Services
Regions}
\itemize{
\item The metering service supports Amazon ECS and EKS for Flexible
Consumption Pricing (FCP) products using MeterUsage API. Amazon ECS
is supported in all Amazon Web Services Regions that MeterUsage API
is available except for GovCloud.
\item Amazon EKS is supported in the following: us-east-1, us-east-2,
us-west-1, us-west-2, eu-west-1, eu-central-1, eu-west-2, eu-west-3,
eu-north-1, ap-east-1, ap-southeast-1, ap-northeast-1,
ap-southeast-2, ap-northeast-2, ap-south-1, ca-central-1, sa-east-1.

For questions about adding Amazon Web Services Regions for metering,
contact \href{mailto://aws.amazon.com/marketplace/management/contact-us/}{Amazon Web Services Marketplace Seller Operations}.
}
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- marketplacemetering(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=marketplacemetering_batch_meter_usage]{batch_meter_usage} \tab The CustomerIdentifier parameter is scheduled for deprecation on March 31, 2026\cr
\link[=marketplacemetering_meter_usage]{meter_usage} \tab API to emit metering records\cr
\link[=marketplacemetering_register_usage]{register_usage} \tab Paid container software products sold through Amazon Web Services Marketplace must integrate with the Amazon Web Services Marketplace Metering Service and call the RegisterUsage operation for software entitlement and metering\cr
\link[=marketplacemetering_resolve_customer]{resolve_customer} \tab ResolveCustomer is called by a SaaS application during the registration process
}
}

\examples{
\dontrun{
svc <- marketplacemetering()
svc$batch_meter_usage(
  Foo = 123
)
}

}
