% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacemetering_operations.R
\name{marketplacemetering_batch_meter_usage}
\alias{marketplacemetering_batch_meter_usage}
\title{BatchMeterUsage is called from a SaaS application listed on AWS
Marketplace to post metering records for a set of customers}
\usage{
marketplacemetering_batch_meter_usage(UsageRecords, ProductCode)
}
\arguments{
\item{UsageRecords}{[required] The set of \code{UsageRecords} to submit.
\code{\link[=marketplacemetering_batch_meter_usage]{batch_meter_usage}} accepts up
to 25 \code{UsageRecords} at a time.}

\item{ProductCode}{[required] Product code is used to uniquely identify a product in AWS Marketplace.
The product code should be the same as the one used during the
publishing of a new product.}
}
\description{
\code{\link[=marketplacemetering_batch_meter_usage]{batch_meter_usage}} is called from a SaaS application listed on AWS Marketplace to post metering records for a set of customers.

See \url{https://paws-r.github.io/docs/marketplacemetering/batch_meter_usage.html} for full documentation.
}
\keyword{internal}
