% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplacemetering_operations.R
\name{marketplacemetering_batch_meter_usage}
\alias{marketplacemetering_batch_meter_usage}
\title{The CustomerIdentifier parameter is scheduled for deprecation on March
31, 2026}
\usage{
marketplacemetering_batch_meter_usage(UsageRecords, ProductCode)
}
\arguments{
\item{UsageRecords}{[required] The set of \code{UsageRecords} to submit.
\code{\link[=marketplacemetering_batch_meter_usage]{batch_meter_usage}} accepts up
to 25 \code{UsageRecords} at a time.}

\item{ProductCode}{[required] Product code is used to uniquely identify a product in Amazon Web
Services Marketplace. The product code should be the same as the one
used during the publishing of a new product.}
}
\description{
The \code{CustomerIdentifier} parameter is scheduled for deprecation on March 31, 2026. Use \code{CustomerAWSAccountID} instead.

See \url{https://www.paws-r-sdk.com/docs/marketplacemetering_batch_meter_usage/} for full documentation.
}
\keyword{internal}
