% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_delete_key}
\alias{paymentcryptographycontrolplane_delete_key}
\title{Deletes the key material and metadata associated with Amazon Web
Services Payment Cryptography key}
\usage{
paymentcryptographycontrolplane_delete_key(
  KeyIdentifier,
  DeleteKeyInDays = NULL
)
}
\arguments{
\item{KeyIdentifier}{[required] The \code{KeyARN} of the key that is scheduled for deletion.}

\item{DeleteKeyInDays}{The waiting period for key deletion. The default value is seven days.}
}
\description{
Deletes the key material and metadata associated with Amazon Web Services Payment Cryptography key.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_delete_key/} for full documentation.
}
\keyword{internal}
