% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_get_parameters_for_import}
\alias{paymentcryptographycontrolplane_get_parameters_for_import}
\title{Gets the import token and the wrapping key certificate to initiate a
TR-34 key import into Amazon Web Services Payment Cryptography}
\usage{
paymentcryptographycontrolplane_get_parameters_for_import(
  KeyMaterialType,
  WrappingKeyAlgorithm
)
}
\arguments{
\item{KeyMaterialType}{[required] The key block format type such as TR-34 or TR-31 to use during key
material import. Import token is only required for TR-34 key import
\code{TR34_KEY_BLOCK}. Import token is not required for TR-31 key import.}

\item{WrappingKeyAlgorithm}{[required] The wrapping key algorithm to generate a wrapping key certificate. This
certificate wraps the key under import within the TR-34 key block
cryptogram. \code{RSA_2048} is the only wrapping key algorithm allowed.}
}
\description{
Gets the import token and the wrapping key certificate to initiate a TR-34 key import into Amazon Web Services Payment Cryptography.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_get_parameters_for_import/} for full documentation.
}
\keyword{internal}
