% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographycontrolplane_operations.R
\name{paymentcryptographycontrolplane_import_key}
\alias{paymentcryptographycontrolplane_import_key}
\title{Imports keys and public key certificates into Amazon Web Services
Payment Cryptography}
\usage{
paymentcryptographycontrolplane_import_key(
  Enabled = NULL,
  KeyCheckValueAlgorithm = NULL,
  KeyMaterial,
  Tags = NULL
)
}
\arguments{
\item{Enabled}{Specifies whether import key is enabled.}

\item{KeyCheckValueAlgorithm}{The algorithm that Amazon Web Services Payment Cryptography uses to
calculate the key check value (KCV) for DES and AES keys.

For DES key, the KCV is computed by encrypting 8 bytes, each with value
'00', with the key to be checked and retaining the 3 highest order bytes
of the encrypted result. For AES key, the KCV is computed by encrypting
8 bytes, each with value '01', with the key to be checked and retaining
the 3 highest order bytes of the encrypted result.}

\item{KeyMaterial}{[required] The key or public key certificate type to use during key material
import, for example TR-34 or RootCertificatePublicKey.}

\item{Tags}{The tags to attach to the key. Each tag consists of a tag key and a tag
value. Both the tag key and the tag value are required, but the tag
value can be an empty (null) string. You can't have more than one tag on
an Amazon Web Services Payment Cryptography key with the same tag key.

You can't have more than one tag on an Amazon Web Services Payment
Cryptography key with the same tag key. If you specify an existing tag
key with a different tag value, Amazon Web Services Payment Cryptography
replaces the current tag value with the specified one.

To use this parameter, you must have
\code{\link[=paymentcryptographycontrolplane_tag_resource]{tag_resource}}
permission.

Don't include confidential or sensitive information in this field. This
field may be displayed in plaintext in CloudTrail logs and other output.

Tagging or untagging an Amazon Web Services Payment Cryptography key can
allow or deny permission to the key.}
}
\description{
Imports keys and public key certificates into Amazon Web Services Payment Cryptography.

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographycontrolplane_import_key/} for full documentation.
}
\keyword{internal}
