% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paymentcryptographydataplane_operations.R
\name{paymentcryptographydataplane_generate_as_2805_kek_validation}
\alias{paymentcryptographydataplane_generate_as_2805_kek_validation}
\title{Establishes node-to-node initialization between payment processing nodes
such as an acquirer, issuer or payment network using Australian Standard
2805 (AS2805)}
\usage{
paymentcryptographydataplane_generate_as_2805_kek_validation(
  KeyIdentifier,
  KekValidationType,
  RandomKeySendVariantMask
)
}
\arguments{
\item{KeyIdentifier}{[required] The \code{keyARN} of sending KEK that Amazon Web Services Payment
Cryptography uses for node-to-node initialization}

\item{KekValidationType}{[required] Parameter information for generating a random key for KEK validation to
perform node-to-node initialization.}

\item{RandomKeySendVariantMask}{[required] The key variant to use for generating a random key for KEK validation
during node-to-node initialization.}
}
\description{
Establishes node-to-node initialization between payment processing nodes such as an acquirer, issuer or payment network using Australian Standard 2805 (AS2805).

See \url{https://www.paws-r-sdk.com/docs/paymentcryptographydataplane_generate_as_2805_kek_validation/} for full documentation.
}
\keyword{internal}
